Balance Statistics for twang Objects

Generates balance statistics for ps, mnps, and iptw objects from twang and for ps.cont objects from WeightIt.

# S3 method for ps
bal.tab(x, 
    stop.method, 
    stats,
    int = FALSE, 
    poly = 1, 
    distance = NULL, 
    addl = NULL, 
    data = NULL, 
    continuous,  
    binary, 
    s.d.denom, 
    thresholds = NULL, 
    weights = NULL,
    cluster = NULL, 
    imp = NULL,
    pairwise = TRUE,
    abs = FALSE,
    subset = NULL,
    quick = TRUE, ...)

# S3 method for mnps
bal.tab(x, 
    stop.method, 
    stats,
    int = FALSE, 
    poly = 1, 
    distance = NULL, 
    addl = NULL, 
    data = NULL, 
    continuous,  
    binary, 
    s.d.denom, 
    thresholds = NULL, 
    weights = NULL,
    cluster = NULL, 
    imp = NULL,
    pairwise = TRUE, 
    abs = FALSE,
    subset = NULL,
    quick = TRUE, ...)
    
# S3 method for iptw
bal.tab(x, 
    stop.method, 
    stats,
    int = FALSE, 
    poly = 1, 
    distance.list = NULL, 
    addl.list = NULL, 
    data = NULL, 
    continuous,  
    binary, 
    s.d.denom, 
    thresholds = NULL, 
    weights = NULL,
    cluster = NULL,
    imp = NULL,
    pairwise = TRUE, 
    abs = FALSE, 
    subset = NULL, 
    quick = TRUE, ...)

# S3 method for ps.cont
bal.tab(x, 
    stop.method, 
    stats,
    int = FALSE, 
    poly = 1, 
    distance = NULL, 
    addl = NULL, 
    data = NULL, 
    continuous,
    binary,
    s.d.denom,
    thresholds = NULL, 
    weights = NULL,
    cluster = NULL, 
    imp = NULL,
    abs = FALSE, 
    subset = NULL, 
    quick = TRUE, ...)

Arguments

x

a ps, mnps, iptw, or ps.cont object; the output of a call to twang::ps(), twang::mnps(), twang::iptw() or WeightIt::ps.cont().

stop.method

a string containing the names of the stopping methods used in the original call to ps(), mnps(), iptw(), or ps.cont(). Examples include "es.max" or "ks.mean" for ps and mnps objects and "p.mean" or "s.max" for ps.cont objects. bal.tab will assess balance for the weights created by those stopping methods. The names can be abbreviated as long as the abbreviations are specific enough. If no stopping methods are provided, bal.tab will default to displaying balance for all available stopping methods.

stats

character; which statistic(s) should be reported. See stats for allowable options. For binary and multi-category treatments, "mean.diffs" (i.e., mean differences) is the default. For continuous treatments, "correlations" (i.e., treatment-covariate Pearson correlations) is the default. Multiple options are allowed.

int

logical or numeric; whether or not to include 2-way interactions of covariates included in covs and in addl. If numeric, will be passed to poly as well.

poly

numeric; the highest polynomial of each continuous covariate to display. For example, if 2, squares of each continuous covariate will be displayed (in addition to the covariate itself); if 3, squares and cubes of each continuous covariate will be displayed, etc. If 1, the default, only the base covariate will be displayed. If int is numeric, poly will take on the value of int.

distance, distance.list

an optional formula or data frame containing distance values (e.g., propensity scores) or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look in the argument to data, if specified. Note that the propensity scores generated by ps() and iptw() (but not mnps() or ps.cont()) are automatically included and named "prop.score.stop.method".

addl, addl.list

an optional formula or data frame containing additional covariates for which to present balance or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look first in the argument to data, if specified, and next in the input object. For iptw objects, can be a list of additional covariate values as described above, with one list entry per time period.

data

an optional data frame containing variables that might be named in arguments to distance, addl, cluster, and imp. Can also be mids object, the output of a call to mice() from the mice package, containing multiply imputed data sets. In this case, imp is automatically supplied using the imputation variable created from processing the mids object.

continuous

whether mean differences for continuous covariates should be standardized ("std") or raw ("raw"). Default "std". Abbreviations allowed. This option can be set globally using set.cobalt.options(). This argument is used to set std in col_w_smd() and col_w_corr.

binary

whether mean differences for binary covariates (i.e., difference in proportion) should be standardized ("std") or raw ("raw"). Default "raw". Abbreviations allowed. This option can be set globally using set.cobalt.options(). This argument is used to set std in col_w_smd() and col_w_corr.

s.d.denom

character; how the denominator for standardized mean differences should be calculated, if requested. See col_w_smd() for allowable options. If not specified, for ps objects, bal.tab() will use "treated" if the estimand of the call to ps() is the ATT and "pooled" if the estimand is the ATE; for mnps objects, bal.tab() will use "treated" if treatATT was specified in the original call to mnps and "pooled" otherwise. Use "all" to get the same values computed by bal.table in twang. Abbreviations allowed.

thresholds

a named vector of balance thresholds, where the name corresponds to the statistic (i.e., in stats) that the threshold applies to. For example, to request thresholds on mean differences and variance ratios, one can set thresholds = c(m = .05, v = 2). Requesting a threshold automatically requests the display of that statistic. See Details.

weights

a named list containing additional weights on which to assess balance. Each entry can be a vector of weights, the name of a variable in data that contains weights, or an object with a get.w() method.

cluster

either a vector containing cluster membership for each unit or a string containing the name of the cluster membership variable in data or the ps, mnps, iptw, or ps.cont object. See bal.tab.cluster for details.

imp

either a vector containing imputation indices for each unit or a string containing the name of the imputation index variable in data or the original data set used in the call to ps(), mnps(), iptw(), or ps.cont(). See bal.tab.imp for details. Not necessary if data is a mids object.

pairwise

whether balance should be computed for pairs of treatments or for each treatment against all groups combined. See bal.tab.multi for details. This can also be used with a binary treatment to assess balance with respect to the full sample.

abs

logical; whether displayed balance statistics should be in absolute value or not.

subset

a logical or numeric vector denoting whether each observation should be included or which observations should be included. If logical, it should be the same length as the variables in the original call to ps(), mnps(), or ps.cont(). NAs will be treated as FALSE. This can be used as an alternative to cluster to examine balance on subsets of the data.

quick

logical; if TRUE, will not compute any values that will not be displayed. Set to FALSE if computed values not displayed will be used later.

...

further arguments to control display of output. See display options for details.

Details

bal.tab.ps() generates a list of balance summaries for the ps object given, and functions similarly to twang::bal.table().

The threshold argument controls whether extra columns should be inserted into the Balance table describing whether the balance statistics in question exceeded or were within the threshold. Including these thresholds also creates summary tables tallying the number of variables that exceeded and were within the threshold and displaying the variables with the greatest imbalance on that balance measure.

Value

For binary point treatments, if clusters are not specified, an object of class "bal.tab" containing balance summaries for the ps object. See bal.tab() for details. If clusters are specified, an object of class "bal.tab.cluster" containing balance summaries within each cluster and a summary of balance across clusters. See bal.tab.cluster for details. If mnps() is used with multi-category treatments, an object of class "bal.tab.multi" containing balance summaries for each pairwise treatment comparison and a summary of balance across pairwise comparisons. See bal.tab.multi for details.

Note

The function bal.table in twang performs a similar function. The variances used in the denominator of the standardized mean difference are weighted and computed using survey::svyvar() in twang and are unweighted here (except when s.weights are specified, in which case col_w_sd is used). twang also uses "all" as the default s.d.denom when the estimand is the ATE; the default here is "pooled". For this reason, results may differ slightly between the two packages.

Author

Noah Greifer

See also

Examples

# \donttest{
library(twang); data("lalonde", package = "cobalt")
#> To reproduce results from prior versions of the twang package, please see the version="legacy" option described in the documentation.

## Using ps() for generalized boosted modeling
ps.out <- ps(treat ~ age + educ + married + race +
             nodegree + re74 + re75, data = lalonde, 
             stop.method = c("ks.mean", "es.mean"), 
             estimand = "ATT", verbose = FALSE)
             
bal.tab(ps.out, stop.method = "ks.mean", un = TRUE, 
        m.threshold = .1, disp.ks = TRUE)
#> Call
#>  ps.fast(formula = formula, data = data, params = params, n.trees = nrounds, 
#>     interaction.depth = max_depth, shrinkage = eta, bag.fraction = subsample, 
#>     n.minobsinnode = min_child_weight, perm.test.iters = perm.test.iters, 
#>     print.level = print.level, verbose = verbose, estimand = estimand, 
#>     stop.method = stop.method, sampw = sampw, multinom = multinom, 
#>     ks.exact = ks.exact, version = version, tree_method = tree_method, 
#>     n.keep = n.keep, n.grid = n.grid, keep.data = keep.data)
#> 
#> Balance Measures
#>                 Type Diff.Un  KS.Un Diff.Adj        M.Threshold KS.Adj
#> prop.score  Distance  2.8072 0.8294   0.5712                    0.2164
#> age          Contin. -0.3094 0.1577   0.0538     Balanced, <0.1 0.0980
#> educ         Contin.  0.0550 0.1114  -0.0810     Balanced, <0.1 0.0678
#> married       Binary -0.3236 0.3236   0.0029     Balanced, <0.1 0.0029
#> race_black    Binary  0.6404 0.6404   0.0176     Balanced, <0.1 0.0176
#> race_hispan   Binary -0.0827 0.0827   0.0014     Balanced, <0.1 0.0014
#> race_white    Binary -0.5577 0.5577  -0.0191     Balanced, <0.1 0.0191
#> nodegree      Binary  0.1114 0.1114   0.0637     Balanced, <0.1 0.0637
#> re74         Contin. -0.7211 0.4470   0.1060 Not Balanced, >0.1 0.0591
#> re75         Contin. -0.2903 0.2876   0.1217 Not Balanced, >0.1 0.0941
#> 
#> Balance tally for mean differences
#>                    count
#> Balanced, <0.1         7
#> Not Balanced, >0.1     2
#> 
#> Variable with the greatest mean difference
#>  Variable Diff.Adj        M.Threshold
#>      re75   0.1217 Not Balanced, >0.1
#> 
#> Effective sample sizes
#>            Control Treated
#> Unadjusted   429.      185
#> Adjusted      25.5     185
# }