cobalt: Covariate Balance Tables and Plots

A set of tools for assessing covariate balance in observational studies numerically and graphically. The functions provide integration with the major R packages used for balancing covariates, including MatchIt, WeightIt, twang, CBPS, and many others, and support objects not made using these packages. They support binary, multi-category and continuous treatments, point and longitudinal treatments, and clustered and multiply imputed data.

The main functions of cobalt are the following:

bal.tab() - generate tables of balance statistics before and after matching, weighting, or subclassification

bal.plot() - generate plots to assess balance visually on one covariate at a time

love.plot() - generate plots to summarize and report balance statistics

Other functions include get.w() for extracting weights from objects produced by other packages, col_w_smd() (and friends documented on the same page) for computing (weighted) balance statistics outside of bal.tab(), bal.compute() for computing scalar balance statistics efficiently, and splitfactor() for splitting factor variables in a dataset into dummy variables.

cobalt has several vignettes, which can be accessed using vignette(package = "cobalt") or visiting the website at https://ngreifer.github.io/cobalt/.

See also

Author

Maintainer: Noah Greifer noah.greifer@gmail.com (ORCID)