Balance statistics for CBPS Objects

Generates balance statistics for CBPS and CBMSM objects from the CBPS package.

# S3 method for CBPS
bal.tab(x,
        stats,
        int = FALSE,
        poly = 1,
        distance = NULL,
        addl = NULL,
        data = NULL,
        continuous,
        binary,
        s.d.denom,
        thresholds = NULL,
        weights = NULL,
        cluster = NULL,
        imp = NULL,
        pairwise = TRUE,
        s.weights = NULL,
        abs = FALSE,
        subset = NULL,
        quick = TRUE,
        ...)

# S3 method for CBMSM
bal.tab(x,
        stats,
        int = FALSE,
        poly = 1,
        distance.list = NULL,
        addl.list = NULL,
        data = NULL,
        continuous,
        binary,
        s.d.denom,
        thresholds = NULL,
        weights = NULL,
        cluster = NULL,
        imp = NULL,
        s.weights = NULL,
        abs = FALSE,
        subset = NULL,
        quick = TRUE,
        ...)

Arguments

x
a CBPS or CBMSM object; the output of a call to CBPS() or CBMSM() from the CBPS package.

stats
character; which statistic(s) should be reported. See stats for allowable options. For binary and multi-category treatments, "mean.diffs" (i.e., mean differences) is the default. For continuous treatments, "correlations" (i.e., treatment-covariate Pearson correlations) is the default. Multiple options are allowed.

int
logical or numeric; whether or not to include 2-way interactions of covariates included in covs and in addl. If numeric, will be passed to poly as well. In older versions of cobalt, setting int = TRUE displayed squares of covariates; to replicate this behavior, set int = 2.

poly
numeric; the highest polynomial of each continuous covariate to display. For example, if 2, squares of each continuous covariate will be displayed (in addition to the covariate itself); if 3, squares and cubes of each continuous covariate will be displayed, etc. If 1, the default, only the base covariate will be displayed. If int is numeric, poly will take on the value of int.

distance, distance.list
an optional formula or data frame containing distance values (e.g., propensity scores) or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look first in the argument to data, if specified, and next in the input object. Note that the propensity scores generated by CBPS() and CBMSM() are automatically included and named "prop.score". For CBMSM objects, can be a list of distance values as described above, with one list entry per time period. Each data set must have one row per individual, unlike the data frame in the original call to CBMSM()

addl, addl.list
an optional formula or data frame containing additional covariates for which to present balance or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look first in the argument to data, if specified, and next in the input object. For CBMSM objects, can be a list of additional covariate values as described above, with one list entry per time period. Each data set must have one row per individual, unlike the data frame in the original call to CBMSM().

data
an optional data frame containing variables that might be named in arguments to distance, addl, cluster, and imp. Can also be mids object, the output of a call to mice from the mice package, containing multiply imputed data sets. In this case, imp is automatically supplied using the imputation variable created from processing the mids object.

continuous
whether mean differences for continuous variables should be standardized ("std") or raw ("raw"). Default "std". Abbreviations allowed. This option can be set globally using set.cobalt.options.

binary
whether mean differences for binary variables (i.e., difference in proportion) should be standardized ("std") or raw ("raw"). Default "raw". Abbreviations allowed. This option can be set globally using set.cobalt.options.

s.d.denom
character; how the denominator for standardized mean differences should be calculated, if requested. See col_w_smd for allowable options. If not specified, bal.tab() will use "treated" if the estimand of the call to CBPS() is the ATT and "pooled" if the estimand is the ATE. Abbreviations allowed.

thresholds
a named vector of balance thresholds, where the name corresponds to the statistic (i.e., in stats) that the threshold applies to. For example, to request thresholds on mean differences and variance ratios, one can set thresholds = c(m = .05, v = 2). Requesting a threshold automatically requests the display of that statistic. See stats.

weights
a named list containing additional weights on which to assess balance. Each entry can be a vector of weights, the name of a variable in data that contains weights, or an object with a get.w method.

cluster
either a vector containing cluster membership for each unit or a string containing the name of the cluster membership variable in data or the CBPS object. See bal.tab.cluster for details.

imp
either a vector containing imputation indices for each unit or a string containing the name of the imputation index variable in data or the original data set used in the call to CBPS() or CBMSM(). See bal.tab.imp for details. Not necessary if data is a mids object.

pairwise
when treatment is multi-category, whether balance should be computed for pairs of treatments or for each treatment against all groups combined. See bal.tab.multi for details.

s.weights
Optional; either a vector containing sampling weights for each unit or a string containing the name of the sampling weight variable in data or the CBPS object. If the original call to CBPS() included sampling weights, they should be specified again here to ensure correct computation of balance statistics and unadjusted values. See Details below.

abs
logical; whether displayed balance statistics should be in absolute value or not.

subset
A logical or numeric vector denoting whether each observation should be included or which observations should be included. If logical, it should be the same length as the variables in the original call to CBPS() or CBMSM(). NAs will be treated as FALSE. This can be used as an alternative to cluster to examine balance on subsets of the data.

quick
logical; if TRUE, will not compute any values that will not be displayed. Set to FALSE if computed values not displayed will be used later.

...
Further arguments to control display of output. See display options for details.

Details

bal.tab.CBPS and codebal.tab.CBMSM generates a list of balance summaries for the CBPS or CBMSM object given and functions similarly to balance() in CBPS.

The thresholds argument controls whether extra columns should be inserted into the Balance table describing whether the balance statistics in question exceeded or were within the threshold. Including these thresholds also creates summary tables tallying the number of variables that exceeded and were within the threshold and displaying the variables with the greatest imbalance on that balance measure.

The CBPS object does not return sampling weights even if they are used; rather, the weights returned already have the sampling weights combined within them. Because some of the checks and defaults in bal.tab() rely on patterns in these weights, using sampling weights in CBPS() without specifying them in bal.tab() can lead to incorrect results. If sampling weights are used in CBPS(), it is important that they are specified in bal.tab() as well.

Value

For point treatments, if clusters are not specified, an object of class "bal.tab" containing balance summaries for the CBPS object. See bal.tab for details.

If clusters are specified, an object of class "bal.tab.cluster" containing balance summaries within each cluster and a summary of balance across clusters. See bal.tab.cluster for details.

If CBPS() is used with multi-category treatments, an object of class "bal.tab.multi" containing balance summaries for each pairwise treatment comparison and a summary of balance across pairwise comparisons. See bal.tab.multi for details.

If CBMSM() is used for longitudinal treatments, an object of class "bal.tab.msm" containing balance summaries for each time period and a summary of balance across time periods. See bal.tab.msm for details.

See also

Examples

library(CBPS)

#> Loading required package: MASS

#> Loading required package: nnet

#> Loading required package: numDeriv

#> Loading required package: glmnet

#> Loading required package: Matrix

#> Loaded glmnet 4.0-2

#> CBPS: Covariate Balancing Propensity Score
#> Version: 0.21
#> Authors: Christian Fong [aut, cre],
#>   Marc Ratkovic [aut],
#>   Kosuke Imai [aut],
#>   Chad Hazlett [ctb],
#>   Xiaolin Yang [ctb],
#>   Sida Peng [ctb]

data("lalonde", package = "cobalt")

## Using CBPS() for generating covariate balancing 
## propensity score weights
cbps.out <- CBPS(treat ~ age + educ + married + race +
             nodegree + re74 + re75, data = lalonde)

#> [1] "Finding ATT with T=1 as the treatment.  Set ATT=2 to find ATT with T=0 as the treatment"

bal.tab(cbps.out)

#> Call
#>  CBPS(formula = treat ~ age + educ + married + race + nodegree + 
#>     re74 + re75, data = lalonde)
#> 
#> Balance Measures
#>                 Type Diff.Adj
#> prop.score  Distance  -0.0057
#> age          Contin.  -0.0052
#> educ         Contin.  -0.0017
#> married       Binary  -0.0029
#> race_black    Binary   0.0019
#> race_hispan   Binary  -0.0002
#> race_white    Binary  -0.0017
#> nodegree      Binary   0.0042
#> re74         Contin.  -0.0078
#> re75         Contin.   0.0061
#> 
#> Effective sample sizes
#>            Control Treated
#> Unadjusted  429.       185
#> Adjusted     99.97     185