Balance Statistics for Matching, optmatch, ebal, and designmatch Objects

Generates balance statistics for output objects from Matching, optmatch, ebal, and designmatch.

# S3 method for Match
bal.tab(x, 
    formula = NULL, 
    data = NULL, 
    treat = NULL, 
    covs = NULL, 
    stats,
    int = FALSE, 
    poly = 1, 
    distance = NULL, 
    addl = NULL, 
    continuous,  
    binary,
    s.d.denom, 
    thresholds = NULL,
    weights = NULL,
    cluster = NULL, 
    imp = NULL,
    pairwise = TRUE,
    abs = FALSE,
    subset = NULL,
    quick = TRUE, 
    ...)
    
# S3 method for optmatch
bal.tab(x, ...)

# S3 method for ebalance
bal.tab(x, ...)

# S3 method for designmatch
bal.tab(x, ...)

Arguments

x

either a Match object (the output of a call to Matching::Match() or Matching::Matchby()), an optmatch object (the output of a call to optmatch::pairmatch() or optmatch::fullmatch()), an ebalance object (the output of a call to ebal::ebalance() or ebal::ebalance.trim()), or the output of a call to designmatch::bmatch() or related wrapper functions from the designmatch package.

formula

a formula with the treatment variable as the response and the covariates for which balance is to be assessed as the predictors. All named variables must be in data. See Details.

data

Optional; a data.frame containing variables with the names used in formula, treat, weights, distance, addl, cluster, and/or imp if any. Can also be a mids object, the output of a call to mice::mice() from the mice package, containing multiply imputed data sets. In this case, imp is automatically supplied using the imputation variable created from processing the mids object. See Details.

treat

a vector of treatment statuses. See Details.

covs

a data frame of covariate values for which to check balance. See Details.

stats

character; which statistic(s) should be reported. See stats for allowable options. For binary and multi-category treatments, "mean.diffs" (i.e., mean differences) is the default. Multiple options are allowed.

int

logical or numeric; whether or not to include 2-way interactions of covariates included in covs and in addl. If numeric, will be passed to poly as well.

poly

numeric; the highest polynomial of each continuous covariate to display. For example, if 2, squares of each continuous covariate will be displayed (in addition to the covariate itself); if 3, squares and cubes of each continuous covariate will be displayed, etc. If 1, the default, only the base covariate will be displayed. If int is numeric, poly will take on the value of int.

distance

an optional formula or data frame containing distance values (e.g., propensity scores) or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look in the argument to data, if specified.

addl

an optional formula or data frame containing additional covariates for which to present balance or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look in the arguments to covs and data, if specified.

continuous

whether mean differences for continuous variables should be standardized ("std") or raw ("raw"). Default "std". Abbreviations allowed. This option can be set globally using set.cobalt.options().

binary

whether mean differences for binary variables (i.e., difference in proportion) should be standardized ("std") or raw ("raw"). Default "raw". Abbreviations allowed. This option can be set globally using set.cobalt.options().

s.d.denom

character; how the denominator for standardized mean differences should be calculated, if requested. See col_w_smd() for allowable options. If not specified, for Match objects, bal.tab() will use "treated" if the estimand of the call to Match() is the ATT, "pooled" if the estimand is the ATE, and "control" if the estimand is the ATC; for optmatch, ebal, and designmatch objects, bal.tab() will determine which value makes the most sense based on the input. Abbreviations allowed.

thresholds

a named vector of balance thresholds, where the name corresponds to the statistic (i.e., in stats) that the threshold applies to. For example, to request thresholds on mean differences and variance ratios, one can set thresholds = c(m = .05, v = 2). Requesting a threshold automatically requests the display of that statistic. See Details.

weights

a named list containing additional weights on which to assess balance. Each entry can be a vector of weights, the name of a variable in data that contains weights, or an object with a get.w() method.

cluster

either a vector containing cluster membership for each unit or a string containing the name of the cluster membership variable in data or the CBPS object. See bal.tab.cluster for details.

imp

either a vector containing imputation indices for each unit or a string containing the name of the imputation index variable in data or the original data set used in the call to weightit(). See bal.tab.imp for details. Not necessary if data is a mids object.

pairwise

whether balance should be computed between the treatment groups or for each treatment against all groups combined. See bal.tab.multi for details.

abs

logical; whether displayed balance statistics should be in absolute value or not.

subset

a logical or numeric vector denoting whether each observation should be included or which observations should be included. If logical, it should be the same length as the variables in the original call to the conditioning function. NAs will be treated as FALSE. This can be used as an alternative to cluster to examine balance on subsets of the data.

quick

logical; if TRUE, will not compute any values that will not be displayed. Set to FALSE if computed values not displayed will be used later.

...

for bal.tab.optmatch(), bal.tab.ebalance(), and bal.tab.designmatch(), the same arguments as those passed to bal.tab.Match(). Otherwise, further arguments to control display of output. See display options for details.

Details

bal.tab() generates a list of balance summaries for the object given, and function similarly to Matching::MatchBalance() and designmatch::meantab(). Note that output objects from designmatch do not have their own class; bal.tab() first checks whether the object meets the criteria to be treated as a designmatch object before dispatching the correct method. Renaming or removing items from the output object can create unintended consequences.

The input to bal.tab.Match(), bal.tab.optmatch(), bal.tab.ebalance(), and bal.tab.designmatch() must include either both formula and data or both covs and treat. Using the formula + data inputs mirrors how Matching::MatchBalance() is used, and using the covs + treat input mirrors how designmatch::meantab() is used. (Note that to see identical results to meantab(), s.d.denom must be set to "pooled", though this is not recommended.) For optmatch output objects, specifying a treatment is not required.

All balance statistics are calculated whether they are displayed by print() or not, unless quick = TRUE. The threshold argument controls whether extra columns should be inserted into the Balance table describing whether the balance statistics in question exceeded or were within the threshold. Including these thresholds also creates summary tables tallying the number of variables that exceeded and were within the threshold and displaying the variables with the greatest imbalance on that balance measure.

The inputs (if any) to covs must be a data frame; if more than one variable is included, this is straightforward (i.e., because data[,c("v1", "v2")] is already a data frame), but if only one variable is used (e.g., data[,"v1"]), R will coerce it to a vector, thus making it unfit for input. To avoid this, simply wrap the input to covs in data.frame() or use subset() if only one variable is to be added. Again, when more than one variable is included, the input is general already a data frame and nothing needs to be done.

Value

For point treatments, if clusters and imputations are not specified, an object of class "bal.tab" containing balance summaries for the given object. See bal.tab() for details. If clusters are specified, an object of class "bal.tab.cluster" containing balance summaries within each cluster and a summary of balance across clusters. See bal.tab.cluster for details.

Author

Noah Greifer

See also

Examples

########## Matching ##########

library(Matching); data("lalonde", package = "cobalt")
#> ## 
#> ##  Matching (Version 4.9-11, Build Date: 2021-10-18)
#> ##  See http://sekhon.berkeley.edu/matching for additional documentation.
#> ##  Please cite software as:
#> ##   Jasjeet S. Sekhon. 2011. ``Multivariate and Propensity Score Matching
#> ##   Software with Automated Balance Optimization: The Matching package for R.''
#> ##   Journal of Statistical Software, 42(7): 1-52. 
#> ##

p.score <- glm(treat ~ age + educ + race + 
            married + nodegree + re74 + re75, 
            data = lalonde, family = "binomial")$fitted.values
Match.out <- Match(Tr = lalonde$treat, X = p.score)

## Using formula and data
bal.tab(Match.out, treat ~ age + educ + race + 
        married + nodegree + re74 + re75, data = lalonde)
#> Balance Measures
#>                Type Diff.Adj
#> age         Contin.   0.2106
#> educ        Contin.   0.0201
#> race_black   Binary   0.0054
#> race_hispan  Binary  -0.0051
#> race_white   Binary  -0.0003
#> married      Binary   0.0661
#> nodegree     Binary  -0.0079
#> re74        Contin.  -0.0772
#> re75        Contin.  -0.0127
#> 
#> Sample sizes
#>                      Control Treated
#> All                   429.       185
#> Matched (ESS)          49.17     185
#> Matched (Unweighted)  136.       185
#> Unmatched             293.         0
if (FALSE) { # requireNamespace("optmatch", quietly = TRUE)
########## optmatch ##########

library("optmatch"); data("lalonde", package = "cobalt")

lalonde$prop.score <- glm(treat ~ age + educ + race + 
            married + nodegree + re74 + re75, 
            data = lalonde, family = binomial)$fitted.values
pm <- pairmatch(treat ~ prop.score, data = lalonde)

## Using formula and data
bal.tab(pm, treat ~ age + educ + race + 
        married + nodegree + re74 + re75, data = lalonde,
        distance = "prop.score")
} ########## ebal ##########

library("ebal"); data("lalonde", package = "cobalt")
#> ##
#> ## ebal Package: Implements Entropy Balancing.
#> ## See http://www.stanford.edu/~jhain/ for additional information.
#> 

covariates <- subset(lalonde, select = -c(re78, treat, race))
e.out <- ebalance(lalonde$treat, covariates)
#> Converged within tolerance 

## Using treat and covs
bal.tab(e.out, treat = lalonde$treat, covs = covariates)
#> Balance Measures
#>             Type Diff.Adj
#> age      Contin.       -0
#> educ     Contin.       -0
#> married   Binary       -0
#> nodegree  Binary        0
#> re74     Contin.       -0
#> re75     Contin.       -0
#> 
#> Effective sample sizes
#>            Control Treated
#> Unadjusted  429.       185
#> Adjusted    247.64     185
########## designmatch ##########
# \donttest{
library("designmatch"); data("lalonde", package = "cobalt")
#> Loading required package: lattice
#> Loading required package: slam
#> Loading required package: Rglpk
#> Using the GLPK callable library version 4.65

covariates <- as.matrix(lalonde[c("age", "educ", "re74", "re75")])
treat <- lalonde$treat
dmout <- bmatch(treat,
                total_groups = sum(treat == 1),
                mom = list(covs = covariates,
                           tols = absstddif(covariates, 
                                            treat, .05))
                )
#>   Building the matching problem... 
#>   GLPK optimizer is open... 
#>   Finding the optimal matches... 
#>   Optimal matches found 

## Using treat and covs
bal.tab(dmout, treat = treat, covs = covariates)
#> Balance Measures
#>         Type Diff.Adj
#> age  Contin.   0.0567
#> educ Contin.  -0.0538
#> re74 Contin.  -0.0613
#> re75 Contin.  -0.0501
#> 
#> Sample sizes
#>           Control Treated
#> All           429     185
#> Matched       185     185
#> Unmatched     244       0
# }