Balance Statistics for cem Objects

Generates balance statistics for cem.match objects from cem.

# S3 method for cem.match
bal.tab(x, 
        data, 
        stats,
        int = FALSE, 
        poly = 1, 
        distance = NULL, 
        addl = NULL, 
        continuous,  
        binary, 
        s.d.denom, 
        thresholds = NULL,
        weights = NULL,
        cluster = NULL, 
        imp = NULL,
        pairwise = TRUE,
        abs = FALSE,
        subset = NULL,
        quick = TRUE, 
        ...)

Arguments

x
a cem.match or cem.match.list object; the output of a call to cem() from the cem package.

data
a data frame containing the treatment, covariates, and variables that might be named in arguments to distance, addl, cluster, and imp. Can also be mids object, the output of a call to mice from the mice package, containing multiply imputed data sets. In this case, imp is automatically supplied using the imputation variable created from processing the mids object. An argument to data is required. It must be the same data used in the call to cem or a mids object from which the data supplied to datalist in the cem call originated.

stats
character; which statistic(s) should be reported. See stats for allowable options. For binary and multi-category treatments, "mean.diffs" (i.e., mean differences) is the default. Multiple options are allowed.

int
logical or numeric; whether or not to include 2-way interactions of covariates included in covs and in addl. If numeric, will be passed to poly as well. In older versions of cobalt, setting int = TRUE displayed squares of covariates; to replicate this behavior, set int = 2.

poly
numeric; the highest polynomial of each continuous covariate to display. For example, if 2, squares of each continuous covariate will be displayed (in addition to the covariate itself); if 3, squares and cubes of each continuous covariate will be displayed, etc. If 1, the default, only the base covariate will be displayed. If int is numeric, poly will take on the value of int.

distance
an optional formula or data frame containing distance values (e.g., propensity scores) or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look in the argument to data, if specified.

addl
an optional formula or data frame containing additional covariates for which to present balance or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look in the arguments to covs and data, if specified.

continuous
whether mean differences for continuous variables should be standardized ("std") or raw ("raw"). Default "std". Abbreviations allowed. This option can be set globally using set.cobalt.options. This argument is used to set std in col_w_smd.

binary
whether mean differences for binary variables (i.e., difference in proportion) should be standardized ("std") or raw ("raw"). Default "raw". Abbreviations allowed. This option can be set globally using set.cobalt.options. This argument is used to set std in col_w_smd.

s.d.denom
character; how the denominator for standardized mean differences should be calculated, if requested. See col_w_smd for allowable options. The default is "treated", where the treated group corresponds to the baseline.group in the call to cem. Abbreviations allowed.

thresholds
a named vector of balance thresholds, where the name corresponds to the statistic (i.e., in stats) that the threshold applies to. For example, to request thresholds on mean differences and variance ratios, one can set thresholds = c(m = .05, v = 2). Requesting a threshold automatically requests the display of that statistic. See Details.

weights
a named list containing additional weights on which to assess balance. Each entry can be a vector of weights, the name of a variable in data that contains weights, or an object with a get.w method.

cluster
either a vector containing cluster membership for each unit or a string containing the name of the cluster membership variable in data. See bal.tab.cluster for details.

imp
either a vector containing imputation indices for each unit or a string containing the name of the imputation index variable in data. See bal.tab.imp for details. Not necessary if data is a mids object.

pairwise
when the treatment is multi-category, whether balance should be computed for pairs of treatments or for each treatment against all others combined. See bal.tab.multi for details. Default is TRUE.

abs
logical; whether displayed balance statistics should be in absolute value or not.

subset
a logical or numeric vector denoting whether each observation should be included or which observations should be included. If logical, it should be the same length as the variables in the original call to cem(). NAs will be treated as FALSE. This can be used as an alternative to cluster to examine balance on subsets of the data.

quick
logical; if TRUE, will not compute any values that will not be displayed. Set to FALSE if computed values not displayed will be used later.

...
further arguments to control display of output. See display options for details.

Details

bal.tab.cem.match() generates a list of balance summaries for the cem.match object given, and functions similarly to imbalance() in cem.

The threshold argument controls whether extra columns should be inserted into the Balance table describing whether the balance statistics in question exceeded or were within the threshold. Including these thresholds also creates summary tables tallying the number of variables that exceeded and were within the threshold and displaying the variables with the greatest imbalance on that balance measure.

Value

If clusters and imputations are not specified, an object of class "bal.tab" containing balance summaries for the cem.match object. See bal.tab for details.

If imputations are specified, an object of class "bal.tab.imp" containing balance summaries for each imputation and a summary of balance across imputations. See bal.tab.imp for details.

If cem() is used with multi-category treatments, an object of class "bal.tab.multi" containing balance summaries for each pairwise treatment comparison. See bal.tab.multi for details.

If clusters are specified, an object of class "bal.tab.cluster" containing balance summaries within each cluster and a summary of balance across clusters. See bal.tab.cluster for details.

Author

Noah Greifer

See also

Examples

if (requireNamespace("cem", quietly = TRUE)) {
library(cem); data("lalonde", package = "cobalt")

## Coarsened exact matching
cem.out <- cem("treat", data = lalonde, drop = "re78")

bal.tab(cem.out, data = lalonde, un = TRUE, 
        stats = c("m", "k"))
}