Balance Statistics for Data Sets

Generates balance statistics for unadjusted, matched, weighted, or stratified data using either a data.frame or formula interface.

# S3 method for data.frame
bal.tab(x, 
        treat,
        stats,
        int = FALSE,
        poly = 1,
        distance = NULL,
        addl = NULL,
        data = NULL,
        continuous,
        binary,
        s.d.denom,
        thresholds = NULL,
        weights = NULL,
        cluster = NULL,
        imp = NULL,
        pairwise = TRUE,
        s.weights = NULL,
        abs = FALSE,
        subset = NULL,
        quick = TRUE,
        subclass = NULL,
        match.strata = NULL,
        method,
        estimand = NULL,
        focal = NULL,
        ...)

# S3 method for formula
bal.tab(x, 
        data = NULL,
        stats,
        int = FALSE,
        poly = 1,
        distance = NULL,
        addl = NULL,
        continuous,
        binary,
        s.d.denom,
        thresholds = NULL,
        weights = NULL,
        cluster = NULL,
        imp = NULL,
        pairwise = TRUE,
        s.weights = NULL,
        abs = FALSE,
        subset = NULL,
        quick = TRUE,
        subclass = NULL,
        match.strata = NULL,
        method,
        estimand = NULL,
        focal = NULL,
        ...)

Arguments

x

either a data.frame containing covariate values for each unit or a formula with the treatment variable as the response and the covariates for which balance is to be assessed as the terms. If a formula is supplied, all terms must be present as variable names in data or the global environment.

treat

either a vector containing treatment status values for each unit or a string containing the name of the treatment variable in data. Required for the data.frame method.

stats, int, poly, distance, addl, data, continuous, binary, thresholds, weights, cluster, imp, pairwise, s.weights, abs, subset, quick, ...

see bal.tab() for details. See below for a special note on the s.d.denom argument.

subclass

optional; either a vector containing subclass membership for each unit or a string containing the name of the subclass variable in data.

match.strata

optional; either a vector containing matching stratum membership for each unit or a string containing the name of the matching stratum variable in data. See Details.

method

character; the method of adjustment, if any. If weights are specified, the user can specify either "matching" or "weighting"; "weighting" is the default. If multiple sets of weights are used, each must have a corresponding value for method, but if they are all of the same type, only one value is required. If subclass is specified, "subclassification" is the default. Abbreviations allowed. The only distinction between "matching" and "weighting" is how sample sizes are displayed.

estimand

character; whether the desired estimand is the "ATT", "ATC", or "ATE" for each set of weights. This argument can be used in place of s.d.denom to specify how standardized differences are calculated.

focal

the name of the focal treatment when multi-category treatments are used. See bal.tab.multi for details.

The following argument has a special note when used with data.frame or formula input objects:

s.d.denom

if weights are supplied, each set of weights should have a corresponding entry to s.d.denom; a single entry will be recycled to all sets of weights. If left blank and one of weights, subclass, or match.strata are supplied, bal.tab() will figure out which one is best based on estimand, if given (for ATT, "treated"; for ATC, "control"; otherwise "pooled") and other clues if not.

Details

bal.tab.data.frame() generates a list of balance summaries for the covariates and treatment status values given. bal.tab.formula() does the same but uses a formula interface instead. When the formula interface is used, the formula and data are reshaped into a treatment vector and data.frame of covariates and then simply passed through the data.frame method.

If weights, subclass and match.strata are all NULL, balance information will be presented only for the unadjusted sample.

The argument to match.strata corresponds to a factor vector containing the name or index of each pair/stratum for units conditioned through matching, for example, using the optmatch package. If more than one of weights, subclass, or match.strata are specified, bal.tab() will attempt to figure out which one to apply. Currently only one of these can be applied ta a time. bal.tab() behaves differently depending on whether subclasses are used in conditioning or not. If they are used, bal.tab() creates balance statistics for each subclass and for the sample in aggregate. See bal.tab.subclass for more information.

Multiple sets of weights can be supplied simultaneously by entering a data.frame or a character vector containing the names of weight variables found in data or a list of weights vectors or names. The arguments to method, s.d.denom, and estimand, if any, must be either the same length as the number of sets of weights or of length one, where the sole entry is applied to all sets. When standardized differences are computed for the unadjusted group, they are done using the first entry to s.d.denom or estimand. When only one set of weights is supplied, the output for the adjusted group will simply be called "Adj", but otherwise will be named after each corresponding set of weights. Specifying multiple sets of weights will also add components to other outputs of bal.tab().

Value

For point treatments, if clusters and imputations are not specified, an object of class "bal.tab" containing balance summaries for the specified treatment and covariates. See bal.tab() for details.

If imputations are specified, an object of class "bal.tab.imp" containing balance summaries for each imputation and a summary of balance across imputations. See bal.tab.imp for details.

If multi-category treatments are used, an object of class "bal.tab.multi" containing balance summaries for each pairwise treatment comparison. See bal.tab.multi for details.

If clusters are specified, an object of class "bal.tab.cluster" containing balance summaries within each cluster and a summary of balance across clusters. See bal.tab.cluster for details.

Author

Noah Greifer

See also

Examples

data("lalonde", package = "cobalt")
lalonde$p.score <- glm(treat ~ age + educ + race, data = lalonde, 
                       family = "binomial")$fitted.values
covariates <- subset(lalonde, select = c(age, educ, race))
                     
## Propensity score weighting using IPTW
lalonde$iptw.weights <- ifelse(lalonde$treat==1, 
                               1/lalonde$p.score, 
                               1/(1-lalonde$p.score))

# data frame interface:
bal.tab(covariates, treat = "treat", data = lalonde, 
        weights = "iptw.weights", s.d.denom = "pooled")
#> Balance Measures
#>                Type Diff.Adj
#> age         Contin.  -0.1242
#> educ        Contin.   0.0727
#> race_black   Binary   0.0053
#> race_hispan  Binary  -0.0025
#> race_white   Binary  -0.0029
#> 
#> Effective sample sizes
#>            Control Treated
#> Unadjusted  429.    185.  
#> Adjusted    344.33   65.47

# Formula interface:
bal.tab(treat ~ age + educ + race, data = lalonde, 
        weights = "iptw.weights", s.d.denom = "pooled")
#> Balance Measures
#>                Type Diff.Adj
#> age         Contin.  -0.1242
#> educ        Contin.   0.0727
#> race_black   Binary   0.0053
#> race_hispan  Binary  -0.0025
#> race_white   Binary  -0.0029
#> 
#> Effective sample sizes
#>            Control Treated
#> Unadjusted  429.    185.  
#> Adjusted    344.33   65.47
      
## Propensity score subclassification
lalonde$subclass <- findInterval(lalonde$p.score, 
                        quantile(lalonde$p.score, 
                        (0:6)/6), all.inside = TRUE)

# data frame interface:
bal.tab(covariates, treat = "treat", data = lalonde, 
        subclass = "subclass", disp.subclass = TRUE, 
        s.d.denom = "pooled")
#> Balance by subclass
#>  - - - Subclass 1 - - - 
#>                Type Diff.Adj
#> age         Contin.  -1.2029
#> educ        Contin.  -0.2551
#> race_black   Binary   0.0000
#> race_hispan  Binary   0.0000
#> race_white   Binary   0.0000
#> 
#>  - - - Subclass 2 - - - 
#>                Type Diff.Adj
#> age         Contin.   0.4108
#> educ        Contin.   0.3005
#> race_black   Binary   0.0000
#> race_hispan  Binary   0.0000
#> race_white   Binary   0.0000
#> 
#>  - - - Subclass 3 - - - 
#>                Type Diff.Adj
#> age         Contin.  -0.1400
#> educ        Contin.   0.0295
#> race_black   Binary   0.0000
#> race_hispan  Binary  -0.0833
#> race_white   Binary   0.0833
#> 
#>  - - - Subclass 4 - - - 
#>                Type Diff.Adj
#> age         Contin.   0.2294
#> educ        Contin.  -0.4409
#> race_black   Binary   0.3467
#> race_hispan  Binary  -0.3467
#> race_white   Binary   0.0000
#> 
#>  - - - Subclass 5 - - - 
#>                Type Diff.Adj
#> age         Contin.   0.4675
#> educ        Contin.   0.3427
#> race_black   Binary   0.0000
#> race_hispan  Binary   0.0000
#> race_white   Binary   0.0000
#> 
#>  - - - Subclass 6 - - - 
#>                Type Diff.Adj
#> age         Contin.   0.1293
#> educ        Contin.  -0.0838
#> race_black   Binary   0.0000
#> race_hispan  Binary   0.0000
#> race_white   Binary   0.0000
#> 

# Formula interface:
bal.tab(treat ~ age + educ + race, data = lalonde, 
      subclass = "subclass", disp.subclass = TRUE, 
      s.d.denom = "pooled")
#> Balance by subclass
#>  - - - Subclass 1 - - - 
#>                Type Diff.Adj
#> age         Contin.  -1.2029
#> educ        Contin.  -0.2551
#> race_black   Binary   0.0000
#> race_hispan  Binary   0.0000
#> race_white   Binary   0.0000
#> 
#>  - - - Subclass 2 - - - 
#>                Type Diff.Adj
#> age         Contin.   0.4108
#> educ        Contin.   0.3005
#> race_black   Binary   0.0000
#> race_hispan  Binary   0.0000
#> race_white   Binary   0.0000
#> 
#>  - - - Subclass 3 - - - 
#>                Type Diff.Adj
#> age         Contin.  -0.1400
#> educ        Contin.   0.0295
#> race_black   Binary   0.0000
#> race_hispan  Binary  -0.0833
#> race_white   Binary   0.0833
#> 
#>  - - - Subclass 4 - - - 
#>                Type Diff.Adj
#> age         Contin.   0.2294
#> educ        Contin.  -0.4409
#> race_black   Binary   0.3467
#> race_hispan  Binary  -0.3467
#> race_white   Binary   0.0000
#> 
#>  - - - Subclass 5 - - - 
#>                Type Diff.Adj
#> age         Contin.   0.4675
#> educ        Contin.   0.3427
#> race_black   Binary   0.0000
#> race_hispan  Binary   0.0000
#> race_white   Binary   0.0000
#> 
#>  - - - Subclass 6 - - - 
#>                Type Diff.Adj
#> age         Contin.   0.1293
#> educ        Contin.  -0.0838
#> race_black   Binary   0.0000
#> race_hispan  Binary   0.0000
#> race_white   Binary   0.0000
#>