Balance Statistics for Longitudinal Datasets

Generates balance statistics for data coming from a longitudinal treatment scenario. The primary input is in the form of a list of formulas or data.frames contain the covariates at each time point. bal.tab automatically classifies this list as either a data.frame.list or formula.list, respectively.

# S3 method for data.frame.list
bal.tab(x,
        treat.list = NULL,
        data = NULL,
        weights = NULL,
        stats,
        int = FALSE,
        poly = 1,
        distance.list = NULL,
        addl.list = NULL,
        method,
        continuous,
        binary,
        s.d.denom,
        thresholds = NULL,
        cluster = NULL,
        imp = NULL,
        pairwise = TRUE,
        s.weights = NULL,
        estimand = "ATE",
        abs = FALSE,
        subset = NULL,
        quick = TRUE,
        ...)

# S3 method for formula.list
bal.tab(x,
        data = NULL,
        ...)

Arguments

x
either a list of data frames containing all the covariates to be assessed at each time point or a list of formulas with the treatment for each time period on the left and the covariates for which balance is to be displayed on the right. Covariates to be assessed at multiple points must be included in the entries for each time point. Data must be in the "wide" format, with one row per unit. If a formula list is supplied, an argument to data is required unless all objects in the formulas exist in the environment.

treat.list
treatment status for each unit at each time point. This can be specified as a list or data frame of vectors, each of which contains the treatment status of each individual at each time point, or a list or vector of the names of variables in data that contain treatment at each time point.

data
for bal.tab.data.frame.list: optional; a data frame containing variables with the names used in treat.list, weights, distance.list, and/or addl.list, if any. For bal.tab.formula.list: required; a data frame containing all covariates named in the formulas and variables with the names used in the arguments mentioned above. If all objects in the x formulas are present in the environment, can be omitted. data must be in the "wide" format, with one row per unit. Can also be mids object, the output of a call to mice from the mice package, containing multiply imputed data sets. In this case, imp is automatically supplied using the imputation variable created from processing the mids object.

weights
optional; a vector, list, or data frame containing weights for each unit or a string containing the names of the weights variables in data. These can be weights generated by, e.g., inverse probability weighting. If weights=NULL, balance information will be presented only for the unadjusted sample.

stats
character; which statistic(s) should be reported. See stats for allowable options. For binary and multi-category treatments, "mean.diffs" (i.e., mean differences) is the default. For continuous treatments, "correlations" (i.e., treatment-covariate Pearson correlations) is the default. Multiple options are allowed.

int
logical or numeric; whether or not to include 2-way interactions of covariates included in covs and in addl. If numeric, will be passed to poly as well. In older versions of cobalt, setting int = TRUE displayed squares of covariates; to replicate this behavior, set int = 2.

poly
numeric; the highest polynomial of each continuous covariate to display. For example, if 2, squares of each continuous covariate will be displayed (in addition to the covariate itself); if 3, squares and cubes of each continuous covariate will be displayed, etc. If 1, the default, only the base covariate will be displayed. If int is numeric, poly will take on the value of int.

distance.list
optional; distance values (e.g., propensity scores) for each unit. These can be specified as a list of vectors or data frames containing the distance values (one for each time point), or as a single vector or data frame to be applied at all times points. The vectors or data frames can be replaced with the names of variables in data containing the distance values. If a list is used and some time points are not to have distance values, these can be replaced with NULL in the list.

addl.list
optional; additional covariates for which to present balance. These may be covariates included in the original dataset but not included in x. In general, it makes more sense to include all desired variables in x than in addl.list. The arguments can be entered the same ways as those to distance.list.

method
a character vector containing the method of adjustment, if any. Currently only "weighting" is supported.

continuous
whether mean differences for continuous variables should be standardized ("std") or raw ("raw"). Default "std". Abbreviations allowed. This option can be set globally using set.cobalt.options.

binary
whether mean differences for binary variables (i.e., difference in proportion) should be standardized ("std") or raw ("raw"). Default "raw". Abbreviations allowed. This option can be set globally using set.cobalt.options.

s.d.denom
character; how the denominator for standardized mean differences should be calculated, if requested. See col_w_smd for allowable options. If weights are supplied, each set of weights should have a corresponding entry to s.d.denom. If left blank and weights are supplied, bal.tab() will try to determine whether the ATT, ATC, or ATE is being estimated based on the pattern of weights and supply s.d.denom accordingly. Abbreviations allowed. If left blank, bal.tab() will try to use the estimand argument. It is recommended not to set this argument for longitudinal treatments.

thresholds
a named vector of balance thresholds, where the name corresponds to the statistic (i.e., in stats) that the threshold applies to. For example, to request thresholds on mean differences and variance ratios, one can set thresholds = c(m = .05, v = 2). Requesting a threshold automatically requests the display of that statistic. See Details.

cluster
either a vector containing cluster membership for each unit or a string containing the name of the cluster membership variable in data. See bal.tab.cluster for details.

imp
either a vector containing imputation indices for each unit or a string containing the name of the imputation index variable in data. See bal.tab.imp for details. Not necessary if data is a mids object.

pairwise
when treatment is multi-category, whether balance should be computed for pairs of treatments or for each treatment against all groups combined. See bal.tab.multi for details.

s.weights
optional; either a vector containing sampling weights for each unit or a string containing the name of the sampling weight variable in data. These function like regular weights except that both the adjusted and unadjusted samples will be weighted according to these weights if weights are used.

estimand
the causal estimand of interest. This value is used to set s.d.denom, and should not be changed from "ATE".

abs
logical; whether displayed balance statistics should be in absolute value or not.

subset
A logical or numeric vector denoting whether each observation should be included or which observations should be included. If logical, it should be the same length as the treatment and covariates. NAs will be treated as FALSE. This can be used as an alternative to cluster to examine balance on subsets of the data.

quick
logical; if TRUE, will not compute any values that will not be displayed. Set to FALSE if computed values not displayed will be used later.

...
For bal.tab.formula.list, other arguments to be passed to bal.tab.data.frame.list. Otherwise, further arguments to control display of output. See display options for details.

Details

bal.tab.formula.list() and bal.tab.data.frame.list() generate a list of balance summaries for each time point based on the treatments and covariates provided. All data must be in the "wide" format, with exactly one row per unit and columns representing variables at different time points. See the weightitMSM documentation for an example of how to transform long data into wide data using reshape.

All balance statistics are calculated whether they are displayed by print or not, unless quick = TRUE. The threshold argument controls whether extra columns should be inserted into the Balance table describing whether the balance statistics in question exceeded or were within the threshold. Including these thresholds also creates summary tables tallying the number of variables that exceeded and were within the threshold and displaying the variables with the greatest imbalance on that balance measure.

Multiple sets of weights can be supplied simultaneously by including entering a data frame or a character vector containing the names of weight variables found in data or a list thereof. The arguments to method, s.d.denom, and estimand, if any, must be either the same length as the number of sets of weights or of length one, where the sole entry is applied to all sets. When standardized differences are computed for the unadjusted group, they are done using the first entry to s.d.denom or estimand. When only one set of weights is supplied, the output for the adjusted group will simply be called "Adj", but otherwise will be named after each corresponding set of weights. Specifying multiple sets of weights will also add components to other output of bal.tab().

Value

An object of class bal.tab.msm containing balance summaries at each time point. Each balance summary is its own bal.tab object. See bal.tab.msm for more details.

Currently, clusters and multiply imputed data are not compatible with longitudinal treatments.

See bal.tab base methods for more detailed information on the value of the bal.tab objects produced for each time point.

See also

Examples

data("iptwExWide", package = "twang")
library("cobalt")

## Estimating longitudinal propensity scores and weights
ps1 <- glm(tx1 ~ age + gender + use0,
            data = iptwExWide,
            family = "binomial")$fitted.values
w1 <- ifelse(iptwExWide$tx1 == 1, 1/ps1, 1/(1-ps1))
ps2 <- glm(tx2 ~ age + gender + use0 + tx1 + use1,
            data = iptwExWide,
            family = "binomial")$fitted.values
w2 <- ifelse(iptwExWide$tx2 == 1, 1/ps2, 1/(1-ps2))
ps3 <- glm(tx3 ~ age + gender + use0 + tx1 + use1 + tx2 + use2,
            data = iptwExWide,
            family = "binomial")$fitted.values
w3 <- ifelse(iptwExWide$tx3 == 1, 1/ps3, 1/(1-ps3))

w <- w1*w2*w3

# Formula interface plus addl.list:
bal.tab(list(tx1 ~ use0 + gender,
             tx2 ~ use0 + gender + use1 + tx1,
             tx3 ~ use0 + gender + use1 + tx1 + use2 + tx2),
        data = iptwExWide,
        weights = w,
        distance.list = list(~ps1, ~ps2, ~ps3),
        addl.list = ~age*gender,
        un = TRUE)

#> Balance summary across all time points
#>                  Times     Type Max.Diff.Un Max.Diff.Adj
#> ps1                  1 Distance      0.7862       0.0251
#> use0           1, 2, 3  Contin.      0.2668       0.0558
#> gender         1, 2, 3   Binary      0.2945       0.0263
#> age            1, 2, 3  Contin.      0.3799       0.0703
#> age * gender_0 1, 2, 3  Contin.      0.4195       0.1022
#> age * gender_1 1, 2, 3  Contin.      0.7019       0.1164
#> ps2                  2 Distance      0.5288       0.0065
#> use1              2, 3  Contin.      0.1662       0.0316
#> tx1               2, 3   Binary      0.1695       0.0171
#> ps3                  3 Distance      0.6565       0.0229
#> use2                 3  Contin.      0.1087       0.0315
#> tx2                  3   Binary      0.2423       0.0085
#> 
#> Effective sample sizes
#>  - Time 1
#>            Control Treated
#> Unadjusted  294.     706. 
#> Adjusted    185.18   573.6
#>  - Time 2
#>            Control Treated
#> Unadjusted   492.   508.  
#> Adjusted     318.9  264.49
#>  - Time 3
#>            Control Treated
#> Unadjusted  415.     585. 
#> Adjusted    235.67   366.4

# data frame interface:
bal.tab(list(iptwExWide[c("use0", "gender")],
             iptwExWide[c("use0", "gender", "use1", "tx1")],
             iptwExWide[c("use0", "gender", "use1", "tx1", "use2", "tx2")]),
        treat.list = iptwExWide[c("tx1", "tx2", "tx3")],
        weights = w,
        distance.list = list(~ps1, ~ps2, ~ps3),
        un = TRUE)

#> Balance summary across all time points
#>          Times     Type Max.Diff.Un Max.Diff.Adj
#> ps1          1 Distance      0.7862       0.0251
#> use0   1, 2, 3  Contin.      0.2668       0.0558
#> gender 1, 2, 3   Binary      0.2945       0.0263
#> ps2          2 Distance      0.5288       0.0065
#> use1      2, 3  Contin.      0.1662       0.0316
#> tx1       2, 3   Binary      0.1695       0.0171
#> ps3          3 Distance      0.6565       0.0229
#> use2         3  Contin.      0.1087       0.0315
#> tx2          3   Binary      0.2423       0.0085
#> 
#> Effective sample sizes
#>  - Time 1
#>            Control Treated
#> Unadjusted  294.     706. 
#> Adjusted    185.18   573.6
#>  - Time 2
#>            Control Treated
#> Unadjusted   492.   508.  
#> Adjusted     318.9  264.49
#>  - Time 3
#>            Control Treated
#> Unadjusted  415.     585. 
#> Adjusted    235.67   366.4