Balance Statistics for MatchIt Objects

Generates balance statistics for matchit objects from MatchIt.

# S3 method for matchit
bal.tab(x, 
    method,
    stats,
    int = FALSE, 
    poly = 1, 
    distance = NULL, 
    addl = NULL, 
    data = NULL, 
    continuous,  
    binary, 
    s.d.denom, 
    thresholds = NULL,
    weights = NULL,
    cluster = NULL, 
    imp = NULL,
    pairwise = TRUE,
    s.weights = NULL,
    abs = FALSE,
    subset = NULL,
    quick = TRUE, 
    ...)

Arguments

x
a matchit object; the output of a call to MatchIt::matchit().

method
a character vector containing the method of adjustment. Ignored unless subclassification was used in the original call to matchit(). If "weighting", the subclassification weights will be used and subclasses will be ignored. If "subclassification", balance will be assessed using the subclasses (see bal.tab.subclass for details). Abbreviations allowed.

stats
character; which statistic(s) should be reported. See stats for allowable options. For binary and multi-category treatments, "mean.diffs" (i.e., mean differences) is the default. Multiple options are allowed.

int
logical or numeric; whether or not to include 2-way interactions of covariates included in covs and in addl. If numeric, will be passed to poly as well.

poly
numeric; the highest polynomial of each continuous covariate to display. For example, if 2, squares of each continuous covariate will be displayed (in addition to the covariate itself); if 3, squares and cubes of each continuous covariate will be displayed, etc. If 1, the default, only the base covariate will be displayed. If int is numeric, poly will take on the value of int.

distance
an optional formula or data frame containing distance values (e.g., propensity scores) or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look in the argument to data, if specified. Note that the distance measure (e.g., propensity score) generated by matchit() is automatically included and named "distance".

addl
an optional formula or data frame containing additional covariates for which to present balance or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look first in the argument to data, if specified, and next in the matchit object.

data
an optional data frame containing variables that might be named in arguments to distance, addl, cluster, and imp. Can also be mids object, the output of a call to mice() from the mice package, containing multiply imputed data sets. In this case, imp is automatically supplied using the imputation variable created from processing the mids object.

continuous
whether mean differences for continuous variables should be standardized ("std") or raw ("raw"). Default "std". Abbreviations allowed. This option can be set globally using set.cobalt.options(). This argument is used to set std in col_w_smd().

binary
whether mean differences for binary variables (i.e., difference in proportion) should be standardized ("std") or raw ("raw"). Default "raw". Abbreviations allowed. This option can be set globally using set.cobalt.options(). This argument is used to set std in col_w_smd().

s.d.denom
character; how the denominator for standardized mean differences should be calculated, if requested. See col_w_smd() for allowable options. If not specified, bal.tab() will figure out which one is best based on the estimand of the matchit object: if ATT, "treated"; if ATC, "control", otherwise "pooled". Abbreviations allowed.

thresholds
a named vector of balance thresholds, where the name corresponds to the statistic (i.e., in stats) that the threshold applies to. For example, to request thresholds on mean differences and variance ratios, one can set thresholds = c(m = .05, v = 2). Requesting a threshold automatically requests the display of that statistic. See Details.

weights
a named list containing additional weights on which to assess balance. Each entry can be a vector of weights, the name of a variable in data that contains weights, or an object with a get.w() method.

cluster
either a vector containing cluster membership for each unit or a string containing the name of the cluster membership variable in data or the matchit object. See bal.tab.cluster for details.

imp
either a vector containing imputation indices for each unit or a string containing the name of the imputation index variable in data or the original data set used in the call to matchit(). See bal.tab.imp for details. Not necessary if data is a mids object.

pairwise
whether balance should be computed between the treatment groups or for each treatment against all groups combined. See bal.tab.multi for details.

s.weights
optional; either a vector containing sampling weights for each unit or a string containing the name of the sampling weight variable in data. These function like regular weights except that both the adjusted and unadjusted samples will be weighted according to these weights if weights are used. If s.weights was supplied in the call to matchit(), they will automatically be included and do not need be specified again (though there is no harm if they are).

abs
logical; whether displayed balance statistics should be in absolute value or not.

subset
a logical or numeric vector denoting whether each observation should be included or which observations should be included. If logical, it should be the same length as the variables in the original call to matchit(). NAs will be treated as FALSE. This can be used as an alternative to cluster to examine balance on subsets of the data.

quick
logical; if TRUE, will not compute any values that will not be displayed. Set to FALSE if computed values not displayed will be used later.

...
further arguments to control display of output. See display options for details.

Details

bal.tab.matchit() generates a list of balance summaries for the matchit object given, and functions similarly to MatchIt::summary.matchit(). bal.tab() behaves differently depending on whether subclasses are used in conditioning or not. If they are used, bal.tab() creates balance statistics for each subclass and for the sample in aggregate; see bal.tab.subclass for more information.

The threshold argument controls whether extra columns should be inserted into the Balance table describing whether the balance statistics in question exceeded or were within the threshold. Including these thresholds also creates summary tables tallying the number of variables that exceeded and were within the threshold and displaying the variables with the greatest imbalance on that balance measure. When subclassification is used, the extra threshold columns are placed within the balance tables for each subclass as well as in the aggregate balance table, and the summary tables display balance for each subclass.

Value

If subclassification is used and method is set to "subclassification", an object of class "bal.tab.subclass" containing balance summaries within and across subclasses. See bal.tab.subclass for details.

If matching is used and clusters are not specified, an object of class "bal.tab" containing balance summaries for the matchit object. See bal.tab() for details.

If clusters are specified, an object of class "bal.tab.cluster" containing balance summaries within each cluster and a summary of balance across clusters. See bal.tab.cluster for details.

Author

Noah Greifer

See also

Examples

if (requireNamespace("MatchIt", quietly = TRUE)) {
library(MatchIt); data("lalonde", package = "cobalt")

## Nearest Neighbor matching
m.out1 <- matchit(treat ~ age + educ + race + 
                  married + nodegree + re74 + re75, 
                  data = lalonde, method = "nearest")
                  
bal.tab(m.out1, un = TRUE, m.threshold = .1, 
        v.threshold = 2)

## Subclassification
m.out2 <- matchit(treat ~ age + educ + race + 
                  married + nodegree + re74 + re75, 
                  data = lalonde, method = "subclass")
                  
bal.tab(m.out2, disp.subclass = TRUE)
}

#> Call
#>  matchit(formula = treat ~ age + educ + race + married + nodegree + 
#>     re74 + re75, data = lalonde, method = "subclass")
#> 
#> Balance by subclass
#>  - - - Subclass 1 - - - 
#>                 Type Diff.Adj
#> distance    Distance   0.2785
#> age          Contin.  -0.4024
#> educ         Contin.   0.1142
#> race_black    Binary   0.0823
#> race_hispan   Binary   0.1492
#> race_white    Binary  -0.2315
#> married       Binary  -0.2877
#> nodegree      Binary  -0.0003
#> re74         Contin.  -0.5864
#> re75         Contin.  -0.1729
#> 
#>  - - - Subclass 2 - - - 
#>                 Type Diff.Adj
#> distance    Distance   0.1873
#> age          Contin.  -0.7473
#> educ         Contin.   0.1183
#> race_black    Binary   0.0094
#> race_hispan   Binary  -0.0094
#> race_white    Binary   0.0000
#> married       Binary  -0.2473
#> nodegree      Binary  -0.0121
#> re74         Contin.  -0.0352
#> re75         Contin.  -0.0970
#> 
#>  - - - Subclass 3 - - - 
#>                 Type Diff.Adj
#> distance    Distance  -0.0140
#> age          Contin.   0.0524
#> educ         Contin.   0.1372
#> race_black    Binary   0.0000
#> race_hispan   Binary   0.0000
#> race_white    Binary   0.0000
#> married       Binary   0.3550
#> nodegree      Binary   0.2191
#> re74         Contin.  -0.2669
#> re75         Contin.  -0.0970
#> 
#>  - - - Subclass 4 - - - 
#>                 Type Diff.Adj
#> distance    Distance  -0.0003
#> age          Contin.  -0.0499
#> educ         Contin.  -0.1436
#> race_black    Binary   0.0000
#> race_hispan   Binary   0.0000
#> race_white    Binary   0.0000
#> married       Binary  -0.1116
#> nodegree      Binary  -0.0417
#> re74         Contin.  -0.0073
#> re75         Contin.  -0.0801
#> 
#>  - - - Subclass 5 - - - 
#>                 Type Diff.Adj
#> distance    Distance  -0.0224
#> age          Contin.   0.2640
#> educ         Contin.  -0.2977
#> race_black    Binary   0.0000
#> race_hispan   Binary   0.0000
#> race_white    Binary   0.0000
#> married       Binary   0.0000
#> nodegree      Binary   0.0376
#> re74         Contin.   0.0190
#> re75         Contin.   0.1233
#> 
#>  - - - Subclass 6 - - - 
#>                 Type Diff.Adj
#> distance    Distance   0.0143
#> age          Contin.   0.5245
#> educ         Contin.   0.2781
#> race_black    Binary   0.0000
#> race_hispan   Binary   0.0000
#> race_white    Binary   0.0000
#> married       Binary   0.0000
#> nodegree      Binary  -0.1290
#> re74         Contin.  -0.0152
#> re75         Contin.  -0.2407
#>