Balance Statistics for MatchThem Objects

Generates balance statistics for mimids and wimids objects from MatchThem.

Usage

# S3 method for class 'mimids'
bal.tab(
  x,
  stats,
  int = FALSE,
  poly = 1,
  distance = NULL,
  addl = NULL,
  data = NULL,
  continuous,
  binary,
  s.d.denom,
  thresholds = NULL,
  weights = NULL,
  cluster = NULL,
  pairwise = TRUE,
  s.weights = NULL,
  abs = FALSE,
  subset = NULL,
  quick = TRUE,
  ...
)

Arguments

x

a mimids or wimids object; the output of a call to MatchThem::matchthem() or MatchThem::weightthem().

stats

character; which statistic(s) should be reported. See stats for allowable options. For binary and multi-category treatments, "mean.diffs" (i.e., mean differences) is the default. For continuous treatments, "correlations" (i.e., treatment-covariate Pearson correlations) is the default. Multiple options are allowed.

int

logical or numeric; whether or not to include 2-way interactions of covariates included in covs and in addl. If numeric, will be passed to poly as well.

poly

numeric; the highest polynomial of each continuous covariate to display. For example, if 2, squares of each continuous covariate will be displayed (in addition to the covariate itself); if 3, squares and cubes of each continuous covariate will be displayed, etc. If 1, the default, only the base covariate will be displayed. If int is numeric, poly will take on the value of int.

distance

an optional formula or data frame containing distance values (e.g., propensity scores) or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look in the argument to data, if specified. The distance measure generated by matchthem() or weightthem() is automatically included and named "distance" or "prop.score", respectively.

addl

an optional formula or data frame containing additional covariates for which to present balance or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look in the arguments to the input object, covs, and data, if specified. For longitudinal treatments, can be a list of allowable arguments, one for each time point.

data

an optional data frame containing variables named in other arguments. For some input object types, this is required.

continuous

whether mean differences for continuous variables should be standardized ("std") or raw ("raw"). Default "std". Abbreviations allowed. This option can be set globally using set.cobalt.options().

binary

whether mean differences for binary variables (i.e., difference in proportion) should be standardized ("std") or raw ("raw"). Default "raw". Abbreviations allowed. This option can be set globally using set.cobalt.options().

s.d.denom

character; how the denominator for standardized mean differences should be calculated, if requested. See col_w_smd() for allowable options. Abbreviations allowed. If not specified, the defaults depend on the options specified in the original function calls; see bal.tab.matchit() and bal.tab.weightit() for details on the defaults.

thresholds

a named vector of balance thresholds, where the name corresponds to the statistic (i.e., in stats) that the threshold applies to. For example, to request thresholds on mean differences and variance ratios, one can set thresholds = c(m = .05, v = 2). Requesting a threshold automatically requests the display of that statistic. When specified, extra columns are inserted into the Balance table describing whether the requested balance statistics exceeded the threshold or not. Summary tables tallying the number of variables that exceeded and were within the threshold and displaying the variables with the greatest imbalance on that balance measure are added to the output.

weights

a vector, list, or data.frame containing weights for each unit, or a string containing the names of the weights variables in data, or an object with a get.w() method or a list thereof. The weights can be, e.g., inverse probability weights or matching weights resulting from a matching algorithm.

cluster

either a vector containing cluster membership for each unit or a string containing the name of the cluster membership variable in data or the input object. See class-bal.tab.cluster for details.

pairwise

whether balance should be computed for pairs of treatments or for each treatment against all groups combined. See bal.tab.multi() for details. This can also be used with a binary treatment to assess balance with respect to the full sample.

s.weights

Optional; either a vector containing sampling weights for each unit or a string containing the name of the sampling weight variable in data. These function like regular weights except that both the adjusted and unadjusted samples will be weighted according to these weights if weights are used.

abs

logical; whether displayed balance statistics should be in absolute value or not.

subset

a logical or numeric vector denoting whether each observation should be included or which observations should be included. If logical, it should have length equal to the number of units. NAs will be treated as FALSE. This can be used as an alternative to cluster to examine balance on subsets of the data.

quick

logical; if TRUE, will not compute any values that will not be displayed. Set to FALSE if computed values not displayed will be used later.

...

for some input types, other arguments that are required or allowed. Otherwise, further arguments to control display of output. See display options for details.

Value

If clusters are not specified, an object of class "bal.tab.imp" containing balance summaries for each imputation and a summary of balance across imputations. See class-bal.tab.imp for details.

If clusters are specified, an object of class "bal.tab.imp.cluster" containing summaries between and across all clusters and imputations.

Details

bal.tab.mimids() and bal.tab.wimids() generate a list of balance summaries for the mimids or wimids object given.

See also

Examples

library(MatchThem)
#> 
#> Attaching package: ‘MatchThem’
#> The following object is masked from ‘package:base’:
#> 
#>     cbind

data("lalonde_mis", package = "cobalt")

#Imputing the missing data
imp <- mice::mice(lalonde_mis, m = 5,
                  print = FALSE)

#Matching using within-imputation propensity scores
mt.out1 <- matchthem(treat ~ age + educ + race + 
                         married + nodegree + re74 + re75, 
                     data = imp, approach = "within")
#> 
#> Matching Observations  | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 
bal.tab(mt.out1)
#> Balance summary across all imputations
#>                 Type Min.Diff.Adj Mean.Diff.Adj Max.Diff.Adj
#> distance    Distance       0.9645        0.9707       0.9753
#> age          Contin.       0.0431        0.0642       0.1224
#> educ         Contin.      -0.1264       -0.1043      -0.0780
#> race_black    Binary       0.3730        0.3730       0.3730
#> race_hispan   Binary      -0.1730       -0.1676      -0.1622
#> race_white    Binary      -0.2108       -0.2054      -0.2000
#> married       Binary      -0.0162       -0.0119      -0.0108
#> nodegree      Binary       0.0541        0.0595       0.0649
#> re74         Contin.      -0.0770       -0.0610      -0.0366
#> re75         Contin.      -0.0686       -0.0500      -0.0301
#> 
#> Average sample sizes across imputations
#>             0   1
#> All       429 185
#> Matched   185 185
#> Unmatched 244   0

#Matching using across-imputation average propensity scores
mt.out2 <- matchthem(treat ~ age + educ + race + 
                         married + nodegree + re74 + re75, 
                     data = imp, approach = "across")
#> Estimating distances   | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 
#> Matching Observations  | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 

bal.tab(mt.out2)
#> Balance summary across all imputations
#>                 Type Min.Diff.Adj Mean.Diff.Adj Max.Diff.Adj
#> distance    Distance       0.9750        0.9750       0.9750
#> age          Contin.       0.0521        0.0521       0.0521
#> educ         Contin.      -0.1129       -0.1129      -0.1129
#> race_black    Binary       0.3730        0.3730       0.3730
#> race_hispan   Binary      -0.1622       -0.1622      -0.1622
#> race_white    Binary      -0.2108       -0.2108      -0.2108
#> married       Binary      -0.0108       -0.0032       0.0054
#> nodegree      Binary       0.0595        0.0595       0.0595
#> re74         Contin.      -0.1287       -0.0862      -0.0572
#> re75         Contin.      -0.0778       -0.0566      -0.0394
#> 
#> Average sample sizes across imputations
#>             0   1
#> All       429 185
#> Matched   185 185
#> Unmatched 244   0

#Weighting using within-imputation propensity scores
wt.out <- weightthem(treat ~ age + educ + race + 
                         married + nodegree + re74 + re75, 
                     data = imp, approach = "within",
                     estimand = "ATT")
#> Estimating weights     | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 

bal.tab(wt.out)
#> Balance summary across all imputations
#>                 Type Min.Diff.Adj Mean.Diff.Adj Max.Diff.Adj
#> prop.score  Distance      -0.0324       -0.0238      -0.0143
#> age          Contin.       0.0980        0.1190       0.1371
#> educ         Contin.      -0.0399       -0.0307      -0.0232
#> race_black    Binary      -0.0033       -0.0023      -0.0014
#> race_hispan   Binary      -0.0004        0.0001       0.0005
#> race_white    Binary       0.0018        0.0022       0.0028
#> married       Binary       0.0122        0.0168       0.0225
#> nodegree      Binary       0.0128        0.0169       0.0228
#> re74         Contin.       0.0013        0.0064       0.0148
#> re75         Contin.      -0.0099       -0.0020       0.0079
#> 
#> Average effective sample sizes across imputations
#>                 0   1
#> Unadjusted 429.   185
#> Adjusted   100.66 185