Balance Statistics for MatchThem Objects

Generates balance statistics for mimids and wimids objects from MatchThem.

Usage

# S3 method for mimids
bal.tab(x, 
        stats,
        int = FALSE,
        poly = 1,
        distance = NULL,
        addl = NULL,
        data = NULL,
        continuous,
        binary,
        s.d.denom,
        thresholds = NULL,
        weights = NULL,
        cluster = NULL,
        pairwise = TRUE,
        s.weights = NULL,
        abs = FALSE,
        subset = NULL,
        quick = TRUE,
        ...)

Arguments

x

a mimids or wimids object; the output of a call to MatchThem::matchthem() or MatchThem::weightthem().

stats, int, poly, addl, data, continuous, binary, thresholds, weights, cluster, pairwise, s.weights, abs, subset, quick, ...

see bal.tab() for details.

See below for special notes on the distance, s.d.denom, and imp arguments. Note the imp argument is ignored because the imputation identifers are already present in the input object.

The following arguments have special notes when used with mimids and wimids objects:

distance

the distance measure generated by matchthem() or weightthem() is automatically included and named "distance" or "prop.score", respectively.

s.d.denom

the defaults depend on the options specified in the original function calls; see bal.tab.matchit() and bal.tab.weightit() for details on the defaults.

Details

bal.tab.mimids() and bal.tab.wimids() generate a list of balance summaries for the mimids or wimids object given.

Value

If clusters are not specified, an object of class "bal.tab.imp" containing balance summaries for each imputation and a summary of balance across imputations. See bal.tab.imp for details.

If clusters are specified, an object of class "bal.tab.imp.cluster" containing summaries between and across all clusters and imputations.

Author

Noah Greifer

See also

Examples

library(mice)
#> 
#> Attaching package: ‘mice’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
#> The following objects are masked from ‘package:base’:
#> 
#>     cbind, rbind
library(MatchThem)
#> 
#> Attaching package: ‘MatchThem’
#> The following objects are masked from ‘package:mice’:
#> 
#>     cbind, pool
#> The following object is masked from ‘package:base’:
#> 
#>     cbind

data("lalonde_mis", package = "cobalt")

#Imputing the missing data
imp <- mice(lalonde_mis, m = 5)
#> 
#>  iter imp variable
#>   1   1  married  re74  re75
#>   1   2  married  re74  re75
#>   1   3  married  re74  re75
#>   1   4  married  re74  re75
#>   1   5  married  re74  re75
#>   2   1  married  re74  re75
#>   2   2  married  re74  re75
#>   2   3  married  re74  re75
#>   2   4  married  re74  re75
#>   2   5  married  re74  re75
#>   3   1  married  re74  re75
#>   3   2  married  re74  re75
#>   3   3  married  re74  re75
#>   3   4  married  re74  re75
#>   3   5  married  re74  re75
#>   4   1  married  re74  re75
#>   4   2  married  re74  re75
#>   4   3  married  re74  re75
#>   4   4  married  re74  re75
#>   4   5  married  re74  re75
#>   5   1  married  re74  re75
#>   5   2  married  re74  re75
#>   5   3  married  re74  re75
#>   5   4  married  re74  re75
#>   5   5  married  re74  re75

#Matching using within-imputation propensity scores
mt.out1 <- matchthem(treat ~ age + educ + race + 
                       married + nodegree + re74 + re75, 
                       data = imp, approach = "within")
#> 
#> Matching Observations  | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 
bal.tab(mt.out1)
#> Balance summary across all imputations
#>                 Type Min.Diff.Adj Mean.Diff.Adj Max.Diff.Adj
#> distance    Distance       0.9675        0.9711       0.9731
#> age          Contin.       0.0416        0.0749       0.1050
#> educ         Contin.      -0.1317       -0.0973      -0.0726
#> race_black    Binary       0.3730        0.3730       0.3730
#> race_hispan   Binary      -0.1784       -0.1665      -0.1568
#> race_white    Binary      -0.2162       -0.2065      -0.1946
#> married       Binary      -0.0270       -0.0119       0.0000
#> nodegree      Binary       0.0432        0.0595       0.0703
#> re74         Contin.      -0.0730       -0.0617      -0.0550
#> re75         Contin.      -0.0774       -0.0504      -0.0217
#> 
#> Average sample sizes across imputations
#>             0   1
#> All       429 185
#> Matched   185 185
#> Unmatched 244   0

#Matching using across-imputation average propensity scores
mt.out2 <- matchthem(treat ~ age + educ + race + 
                       married + nodegree + re74 + re75, 
                       data = imp, approach = "across")
#> Estimating distances   | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 
#> Matching Observations  | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 
                       
bal.tab(mt.out2)
#> Balance summary across all imputations
#>                 Type Min.Diff.Adj Mean.Diff.Adj Max.Diff.Adj
#> distance    Distance       0.9746        0.9746       0.9746
#> age          Contin.       0.0778        0.0778       0.0778
#> educ         Contin.      -0.0968       -0.0968      -0.0968
#> race_black    Binary       0.3730        0.3730       0.3730
#> race_hispan   Binary      -0.1730       -0.1730      -0.1730
#> race_white    Binary      -0.2000       -0.2000      -0.2000
#> married       Binary      -0.0270       -0.0130      -0.0054
#> nodegree      Binary       0.0595        0.0595       0.0595
#> re74         Contin.      -0.1033       -0.0917      -0.0790
#> re75         Contin.      -0.0916       -0.0657      -0.0463
#> 
#> Average sample sizes across imputations
#>             0   1
#> All       429 185
#> Matched   185 185
#> Unmatched 244   0

#Weighting using within-imputation propensity scores
wt.out <- weightthem(treat ~ age + educ + race + 
                       married + nodegree + re74 + re75, 
                       data = imp, approach = "within",
                       estimand = "ATT")
#> Estimating weights     | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 
                       
bal.tab(wt.out)
#> Balance summary across all imputations
#>                 Type Min.Diff.Adj Mean.Diff.Adj Max.Diff.Adj
#> prop.score  Distance      -0.0275       -0.0239      -0.0203
#> age          Contin.       0.1022        0.1119       0.1286
#> educ         Contin.      -0.0392       -0.0341      -0.0282
#> race_black    Binary      -0.0030       -0.0025      -0.0023
#> race_hispan   Binary       0.0001        0.0002       0.0003
#> race_white    Binary       0.0020        0.0023       0.0027
#> married       Binary       0.0148        0.0169       0.0192
#> nodegree      Binary       0.0161        0.0189       0.0212
#> re74         Contin.      -0.0000        0.0062       0.0128
#> re75         Contin.      -0.0059        0.0074       0.0221
#> 
#> Average effective sample sizes across imputations
#>                 0   1
#> Unadjusted 429.   185
#> Adjusted   100.45 185