Balance Statistics for MatchThem Objects

Generates balance statistics for mimids and wimids objects from MatchThem.

# S3 method for mimids
bal.tab(x, 
    stats,
    int = FALSE, 
    poly = 1, 
    distance = NULL, 
    addl = NULL, 
    data = NULL, 
    continuous,  
    binary, 
    s.d.denom, 
    thresholds = NULL,
    weights = NULL,
    cluster = NULL, 
    pairwise = TRUE,
    abs = FALSE,
    subset = NULL,
    quick = TRUE, 
    ...)

# S3 method for wimids
bal.tab(x, 
    stats,
    int = FALSE, 
    poly = 1, 
    distance = NULL, 
    addl = NULL, 
    data = NULL, 
    continuous,  
    binary, 
    s.d.denom, 
    thresholds = NULL,
    weights = NULL,
    cluster = NULL, 
    pairwise = TRUE, 
    abs = FALSE,
    subset = NULL,
    quick = TRUE, 
    ...)

Arguments

x

a mimids or wimids object; the output of a call to MatchThem::matchthem() or MatchThem::weightthem().

stats

character; which statistic(s) should be reported. See stats for allowable options. For binary and multi-category treatments, "mean.diffs" (i.e., mean differences) is the default. For continuous treatments, "correlations" (i.e., treatment-covariate Pearson correlations) is the default. Multiple options are allowed.

int

logical or numeric; whether or not to include 2-way interactions of covariates included in covs and in addl. If numeric, will be passed to poly as well.

poly

numeric; the highest polynomial of each continuous covariate to display. For example, if 2, squares of each continuous covariate will be displayed (in addition to the covariate itself); if 3, squares and cubes of each continuous covariate will be displayed, etc. If 1, the default, only the base covariate will be displayed. If int is numeric, poly will take on the value of int.

distance

an optional formula or data frame containing distance values (e.g., propensity scores) or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look in the argument to data, if specified. Note that the distance measure generated by matchthem() or weightthem() is automatically included and named "distance" or "prop.score", respectively.

addl

an optional formula or data frame containing additional covariates for which to present balance or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look first in the argument to data, if specified, and next in the mimids or wimids object.

data

an optional data frame or mids object containing variables that might be named in arguments to distance, addl, and cluster. See Examples.

continuous

whether mean differences for continuous variables should be standardized ("std") or raw ("raw"). Default "std". Abbreviations allowed. This option can be set globally using set.cobalt.options().

binary

whether mean differences for binary variables (i.e., difference in proportion) should be standardized ("std") or raw ("raw"). Default "raw". Abbreviations allowed. This option can be set globally using set.cobalt.options().

s.d.denom

character; how the denominator for standardized mean differences should be calculated, if requested. See col_w_smd() for allowable options. The defaults depend on the options specified in the original function calls; see bal.tab.matchit() and bal.tab.weightit() for details on the defaults. Abbreviations allowed.

thresholds

a named vector of balance thresholds, where the name corresponds to the statistic (i.e., in stats) that the threshold applies to. For example, to request thresholds on mean differences and variance ratios, one can set thresholds = c(m = .05, v = 2). Requesting a threshold automatically requests the display of that statistic. See Details.

weights

a named list containing additional weights on which to assess balance. Each entry can be a vector of weights, the name of a variable in data that contains weights, or an object with a get.w() method.

cluster

either a vector containing cluster membership for each unit or a string containing the name of the cluster membership variable in data or the mimids or wimids object. See bal.tab.cluster for details.

pairwise

whether balance should be computed for pairs of treatments or for each treatment against all groups combined. See bal.tab.multi for details. This can also be used with a binary treatment to assess balance with respect to the full sample.

abs

logical; whether displayed balance statistics should be in absolute value or not.

subset

a logical or numeric vector denoting whether each observation should be included or which observations should be included. If logical, it should be the same length as the variables in the original (unimputed) dataset or the call to matchthem() or weightthem() (i.e., one for each individual or one for each individual for each imputation). NAs will be treated as FALSE. This can be used as an alternative to cluster to examine balance on subsets of the data.

quick

logical; if TRUE, will not compute any values that will not be displayed. Set to FALSE if computed values not displayed will be used later.

...

further arguments to control display of output. See display options for details.

Details

bal.tab.mimids() and bal.tab.wimids() generate a list of balance summaries for the mimids or wimids object given.

The threshold argument controls whether extra columns should be inserted into the Balance table describing whether the balance statistics in question exceeded or were within the threshold. Including these thresholds also creates summary tables tallying the number of variables that exceeded and were within the threshold and displaying the variables with the greatest imbalance on that balance measure.

Value

If clusters are not specified, an object of class "bal.tab.imp" containing balance summaries for each imputation and a summary of balance across imputations. See bal.tab.imp for details. If clusters are specified, an object of class "bal.tab.imp.cluster" containing summaries between and across all clusters and imputations.

Author

Noah Greifer

See also

Examples

if (all(sapply(c("mice", "MatchThem"), requireNamespace, quietly = TRUE))) {
library(mice)
library(MatchThem)

data("lalonde_mis", package = "cobalt")

#Imputing the missing data
imp <- mice(lalonde_mis, m = 5)

#Matching using within-imputation propensity scores
mt.out1 <- matchthem(treat ~ age + educ + race + 
                       married + nodegree + re74 + re75, 
                       data = imp, approach = "within")
bal.tab(mt.out1)

#Matching using across-imputation average propensity scores
mt.out2 <- matchthem(treat ~ age + educ + race + 
                       married + nodegree + re74 + re75, 
                       data = imp, approach = "across")
                       
bal.tab(mt.out2)

#Weighting using within-imputation propensity scores
wt.out <- weightthem(treat ~ age + educ + race + 
                       married + nodegree + re74 + re75, 
                       data = imp, approach = "within",
                       estimand = "ATT")
                       
bal.tab(wt.out)
}
#> 
#> Attaching package: ‘mice’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
#> The following objects are masked from ‘package:base’:
#> 
#>     cbind, rbind
#> 
#> Attaching package: ‘MatchThem’
#> The following objects are masked from ‘package:mice’:
#> 
#>     cbind, pool
#> The following object is masked from ‘package:base’:
#> 
#>     cbind
#> 
#>  iter imp variable
#>   1   1  married  re74  re75
#>   1   2  married  re74  re75
#>   1   3  married  re74  re75
#>   1   4  married  re74  re75
#>   1   5  married  re74  re75
#>   2   1  married  re74  re75
#>   2   2  married  re74  re75
#>   2   3  married  re74  re75
#>   2   4  married  re74  re75
#>   2   5  married  re74  re75
#>   3   1  married  re74  re75
#>   3   2  married  re74  re75
#>   3   3  married  re74  re75
#>   3   4  married  re74  re75
#>   3   5  married  re74  re75
#>   4   1  married  re74  re75
#>   4   2  married  re74  re75
#>   4   3  married  re74  re75
#>   4   4  married  re74  re75
#>   4   5  married  re74  re75
#>   5   1  married  re74  re75
#>   5   2  married  re74  re75
#>   5   3  married  re74  re75
#>   5   4  married  re74  re75
#>   5   5  married  re74  re75
#> 
#> Matching Observations  | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 
#> Estimating distances   | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 
#> Matching Observations  | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 
#> Estimating weights     | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 
#> Balance summary across all imputations
#>                 Type Min.Diff.Adj Mean.Diff.Adj Max.Diff.Adj
#> prop.score  Distance      -0.0404       -0.0275      -0.0180
#> age          Contin.       0.0853        0.1029       0.1287
#> educ         Contin.      -0.0330       -0.0306      -0.0264
#> race_black    Binary      -0.0041       -0.0027      -0.0015
#> race_hispan   Binary      -0.0006        0.0003       0.0009
#> race_white    Binary       0.0016        0.0023       0.0032
#> married       Binary       0.0140        0.0180       0.0247
#> nodegree      Binary       0.0145        0.0172       0.0192
#> re74         Contin.       0.0095        0.0159       0.0241
#> re75         Contin.      -0.0072        0.0056       0.0151
#> 
#> Average effective sample sizes across imputations
#>            Control Treated
#> Unadjusted  429.       185
#> Adjusted    100.67     185