Balance Statistics for twang Objects

Generates balance statistics for ps, mnps, and iptw objects from twang and for ps.cont objects from WeightIt. Note that several arguments that used to be documented here are now documented in display options. They are still available.

# S3 method for ps
bal.tab(x,
    stop.method,
    stats,
    int = FALSE,
    poly = 1,
    distance = NULL,
    addl = NULL,
    data = NULL,
    continuous,
    binary,
    s.d.denom,
    thresholds = NULL,
    weights = NULL,
    cluster = NULL,
    imp = NULL,
    abs = FALSE,
    subset = NULL,
    quick = TRUE, ...)

# S3 method for mnps
bal.tab(x,
    stop.method,
    stats,
    int = FALSE,
    poly = 1,
    distance = NULL,
    addl = NULL,
    data = NULL,
    continuous,
    binary,
    s.d.denom,
    thresholds = NULL,
    weights = NULL,
    cluster = NULL,
    imp = NULL,
    pairwise = TRUE,
    abs = FALSE,
    subset = NULL,
    quick = TRUE, ...)

# S3 method for iptw
bal.tab(x,
    stop.method,
    stats,
    int = FALSE,
    poly = 1,
    distance.list = NULL,
    addl.list = NULL,
    data = NULL,
    continuous,
    binary,
    s.d.denom,
    thresholds = NULL,
    weights = NULL,
    cluster = NULL,
    imp = NULL,
    pairwise = TRUE,
    abs = FALSE,
    subset = NULL,
    quick = TRUE, ...)

# S3 method for ps.cont
bal.tab(x,
    stop.method,
    stats,
    int = FALSE,
    poly = 1,
    distance = NULL,
    addl = NULL,
    data = NULL,
    continuous,
    binary,
    s.d.denom,
    thresholds = NULL,
    weights = NULL,
    cluster = NULL,
    imp = NULL,
    abs = FALSE,
    subset = NULL,
    quick = TRUE, ...)

Arguments

x
a ps, mnps, iptw, or ps.cont object; the output of a call to ps(), mnps(), or iptw() from twang or from a call to ps.cont() from WeightIt.

stop.method
a string containing the names of the stopping methods used in the original call to ps(), mnps(), iptw(), or ps.cont(). Examples include "es.max" or "ks.mean" for ps and mnps objects and "p.mean" or "s.max" for ps.cont objects. bal.tab will assess balance for the weights created by those stopping methods. The names can be abbreviated as long as the abbreviations are specific enough. If no stopping methods are provided, bal.tab will default to displaying balance for all available stopping methods. This argument used to be called full.stop.method, and that name still works.

stats
character; which statistic(s) should be reported. See stats for allowable options. For binary and multi-category treatments, "mean.diffs" (i.e., mean differences) is the default. For continuous treatments, "correlations" (i.e., treatment-covariate Pearson correlations) is the default. Multiple options are allowed.

int
logical or numeric; whether or not to include 2-way interactions of covariates included in covs and in addl. If numeric, will be passed to poly as well. In older versions of cobalt, setting int = TRUE displayed squares of covariates; to replicate this behavior, set int = 2.

poly
numeric; the highest polynomial of each continuous covariate to display. For example, if 2, squares of each continuous covariate will be displayed (in addition to the covariate itself); if 3, squares and cubes of each continuous covariate will be displayed, etc. If 1, the default, only the base covariate will be displayed. If int is numeric, poly will take on the value of int.

distance, distance.list
an optional formula or data frame containing distance values (e.g., propensity scores) or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look in the argument to data, if specified. Note that the propensity scores generated by ps() and iptw() (but not mnps() or ps.cont()) are automatically included and named "prop.score.stop.method".

addl, addl.list
an optional formula or data frame containing additional covariates for which to present balance or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look first in the argument to data, if specified, and next in the input object. For iptw objects, can be a list of additional covariate values as described above, with one list entry per time period.

data
an optional data frame containing variables that might be named in arguments to distance, addl, cluster, and imp. Can also be mids object, the output of a call to mice from the mice package, containing multiply imputed data sets. In this case, imp is automatically supplied using the imputation variable created from processing the mids object.

continuous
whether mean differences for continuous covariates should be standardized ("std") or raw ("raw"). Default "std". Abbreviations allowed. This option can be set globally using set.cobalt.options. This argument is used to set std in col_w_smd and col_w_corr.

binary
whether mean differences for binary covariates (i.e., difference in proportion) should be standardized ("std") or raw ("raw"). Default "raw". Abbreviations allowed. This option can be set globally using set.cobalt.options. This argument is used to set std in col_w_smd and col_w_corr.

s.d.denom
character; how the denominator for standardized mean differences should be calculated, if requested. See col_w_smd for allowable options. If not specified, for ps objects, bal.tab() will use "treated" if the estimand of the call to ps() is the ATT and "pooled" if the estimand is the ATE; for mnps objects, bal.tab() will use "treated" if treatATT was specified in the original call to mnps and "pooled" otherwise. Use "all" to get the same values computed by bal.table in twang. Abbreviations allowed.

thresholds
a named vector of balance thresholds, where the name corresponds to the statistic (i.e., in stats) that the threshold applies to. For example, to request thresholds on mean differences and variance ratios, one can set thresholds = c(m = .05, v = 2). Requesting a threshold automatically requests the display of that statistic. See Details.

weights
a named list containing additional weights on which to assess balance. Each entry can be a vector of weights, the name of a variable in data that contains weights, or an object with a get.w method.

cluster
either a vector containing cluster membership for each unit or a string containing the name of the cluster membership variable in data or the ps, mnps, iptw, or ps.cont object. See bal.tab.cluster for details.

imp
either a vector containing imputation indices for each unit or a string containing the name of the imputation index variable in data or the original data set used in the call to ps(), mnps(), iptw(), or ps.cont(). See bal.tab.imp for details. Not necessary if data is a mids object.

pairwise
when treatment is multi-category, whether balance should be computed for pairs of treatments or for each treatment against all groups combined. See bal.tab.multi for details.

abs
logical; whether displayed balance statistics should be in absolute value or not.

subset
a logical or numeric vector denoting whether each observation should be included or which observations should be included. If logical, it should be the same length as the variables in the original call to ps(), mnps(), or ps.cont(). NAs will be treated as FALSE. This can be used as an alternative to cluster to examine balance on subsets of the data.

quick
logical; if TRUE, will not compute any values that will not be displayed. Set to FALSE if computed values not displayed will be used later.

...
further arguments to control display of output. See display options for details.

Details

bal.tab.ps() generates a list of balance summaries for the ps object given, and functions similarly to bal.table() in twang.

All balance statistics are calculated whether they are displayed by print or not, unless quick = TRUE. The threshold argument controls whether extra columns should be inserted into the Balance table describing whether the balance statistics in question exceeded or were within the threshold. Including these thresholds also creates summary tables tallying the number of variables that exceeded and were within the threshold and displaying the variables with the greatest imbalance on that balance measure.

Value

For binary point treatments, if clusters are not specified, an object of class "bal.tab" containing balance summaries for the ps object. See bal.tab for details.

If clusters are specified, an object of class "bal.tab.cluster" containing balance summaries within each cluster and a summary of balance across clusters. See bal.tab.cluster for details.

If mnps() is used with multi-category treatments, an object of class "bal.tab.multi" containing balance summaries for each pairwise treatment comparison and a summary of balance across pairwise comparisons. See bal.tab.multi for details.

If ps.cont() is used with continuous treatments, means, mean differences, and variance ratios are replaced by (weighted) Pearson correlations between each covariate and treatment as calculated using col_w_corr. The r.threshold argument works the same as m.threshold, v.threshold, or ks.threshold, adding an extra column to the balance table output and creating additional summaries for balance tallies and maximum imbalances. All arguments related to the calculation or display of mean differences or variance ratios are ignored. The int, addl, un, distance, and cluster arguments are still used as described above.

Note

The function bal.table in twang performs a similar function. The variances used in the denominator of the standardized mean difference are weighted and computed using survey::svyvar in twang and are unweighted here (except when s.weights are specified, in which case col_w_sd is used). twang also uses "all" as the default s.d.denom when the estimand is the ATE; the default here is "pooled". For this reason, results may differ slightly between the two packages.

See also

Examples

library(twang); data("lalonde", package = "cobalt")

#> Loading required package: gbm

#> Loaded gbm 2.1.8

#> Loading required package: survey

#> Loading required package: grid

#> 
#> Attaching package: ‘survey’

#> The following object is masked from ‘package:graphics’:
#> 
#>     dotchart

#> Loading required package: xtable

#> Loading required package: latticeExtra

if (FALSE) {
## Using ps() for generalized boosted modeling
ps.out <- ps(treat ~ age + educ + married + race +
             nodegree + re74 + re75, data = lalonde,
             stop.method = c("ks.mean", "es.mean"),
             estimand = "ATT", verbose = FALSE)
bal.tab(ps.out, stop.method = "es.max", un = TRUE,
        m.threshold = .1, disp.ks = TRUE)
}