Balance Statistics for sbw Objects

Generates balance statistics for sbwcau objects from sbw.

# S3 method for sbwcau
bal.tab(x, 
    stats,
    int = FALSE, 
    poly = 1, 
    distance = NULL, 
    addl = NULL, 
    data = NULL, 
    continuous,  
    binary, 
    s.d.denom, 
    thresholds = NULL,
    weights = NULL,
    cluster = NULL, 
    imp = NULL,
    pairwise = TRUE,
    abs = FALSE,
    subset = NULL,
    quick = TRUE, 
    ...)

Arguments

x

an sbwcau object; the output of a call to sbw::sbw().

stats

character; which statistic(s) should be reported. See stats for allowable options. For binary treatments, "mean.diffs" (i.e., mean differences) is the default. Multiple options are allowed.

int

logical or numeric; whether or not to include 2-way interactions of covariates included in covs and in addl. If numeric, will be passed to poly as well.

poly

numeric; the highest polynomial of each continuous covariate to display. For example, if 2, squares of each continuous covariate will be displayed (in addition to the covariate itself); if 3, squares and cubes of each continuous covariate will be displayed, etc. If 1, the default, only the base covariate will be displayed. If int is numeric, poly will take on the value of int.

distance

an optional formula or data frame containing distance values (e.g., propensity scores) or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look in the argument to data, if specified.

addl

an optional formula or data frame containing distance values (e.g., propensity scores) or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look first in the argument to data, if specified, and next in the sbwcau object.

data

an optional data frame containing variables that might be named in arguments to distance, addl, cluster, and imp. Can also be mids object, the output of a call to mice() from the mice package, containing multiply imputed data sets. In this case, imp is automatically supplied using the imputation variable created from processing the mids object.

continuous

whether mean differences for continuous variables should be standardized ("std") or raw ("raw"). Default "std". Abbreviations allowed. This option can be set globally using set.cobalt.options(). This argument is used to set std in col_w_smd().

binary

whether mean differences for binary variables (i.e., difference in proportion) should be standardized ("std") or raw ("raw"). Default "raw". Abbreviations allowed. This option can be set globally using set.cobalt.options(). This argument is used to set std in col_w_smd().

s.d.denom

character; how the denominator for standardized mean differences should be calculated, if requested. See col_w_smd() for allowable options. If not specified, bal.tab() will figure out which one is best based on the par component of the sbwcau object: if "att", "treated"; if "atc", "control"; otherwise "pooled". Abbreviations allowed.

thresholds

a named vector of balance thresholds, where the name corresponds to the statistic (i.e., in stats) that the threshold applies to. For example, to request thresholds on mean differences and variance ratios, one can set thresholds = c(m = .05, v = 2). Requesting a threshold automatically requests the display of that statistic. See Details.

weights

a named list containing additional weights on which to assess balance. Each entry can be a vector of weights, the name of a variable in data that contains weights, or an object with a get.w() method.

cluster

either a vector containing cluster membership for each unit or a string containing the name of the cluster membership variable in data or the sbwcau object. See bal.tab.cluster for details.

imp

either a vector containing imputation indices for each unit or a string containing the name of the imputation index variable in data or the original data set used in the call to sbw(). See bal.tab.imp for details. Not necessary if data is a mids object.

pairwise

whether balance should be computed between the treatment groups or for each treatment against all groups combined. See bal.tab.multi for details.

abs

logical; whether displayed balance statistics should be in absolute value or not.

subset

a logical or numeric vector denoting whether each observation should be included or which observations should be included. If logical, it should be the same length as the variables in the original call to sbw(). NAs will be treated as FALSE. This can be used as an alternative to cluster to examine balance on subsets of the data.

quick

logical; if TRUE, will not compute any values that will not be displayed. Set to FALSE if computed values not displayed will be used later.

...

further arguments to control display of output. See display options for details.

Details

bal.tab.sbwcau() generates a list of balance summaries for the sbwcau object given, and functions similarly to sbw::summarize().

The thresholds argument controls whether extra columns should be inserted into the Balance table describing whether the balance statistics in question exceeded or were within the threshold. Including these thresholds also creates summary tables tallying the number of variables that exceeded and were within the threshold and displaying the variables with the greatest imbalance on that balance measure.

Value

If clusters are not specified, an object of class "bal.tab" containing balance summaries for the sbwcau object. See bal.tab() for details. If clusters are specified, an object of class "bal.tab.cluster" containing balance summaries within each cluster and a summary of balance across clusters. See bal.tab.cluster for details.

Author

Noah Greifer

See also

Examples