Balance Statistics for WeightIt Objects

Generates balance statistics for weightit and weightitMSM objects from WeightIt.

# S3 method for weightit
bal.tab(x,
        stats,
        int = FALSE,
        poly = 1,
        distance = NULL,
        addl = NULL,
        data = NULL,
        continuous,
        binary,
        s.d.denom,
        thresholds = NULL,
        weights = NULL,
        cluster = NULL,
        imp = NULL,
        pairwise = TRUE,
        abs = FALSE,
        subset = NULL,
        quick = TRUE,
        ...)

# S3 method for weightitMSM
bal.tab(x,
        stats,
        int = FALSE,
        poly = 1,
        distance.list = NULL,
        addl.list = NULL,
        data = NULL,
        continuous,
        binary,
        s.d.denom,
        thresholds = NULL,
        weights = NULL,
        cluster = NULL,
        imp = NULL,
        pairwise = TRUE,
        abs = FALSE,
        subset = NULL,
        quick = TRUE,
        ...)

Arguments

x
a weightit or weightitMSM object; the output of a call to weightit() or weightitMSM() from the WeightIt package.

stats
character; which statistic(s) should be reported. See stats for allowable options. For binary and multi-category treatments, "mean.diffs" (i.e., mean differences) is the default. For continuous treatments, "correlations" (i.e., treatment-covariate Pearson correlations) is the default. Multiple options are allowed.

int
logical or numeric; whether or not to include 2-way interactions of covariates included in covs and in addl. If numeric, will be passed to poly as well. In older versions of cobalt, setting int = TRUE displayed squares of covariates; to replicate this behavior, set int = 2.

poly
numeric; the highest polynomial of each continuous covariate to display. For example, if 2, squares of each continuous covariate will be displayed (in addition to the covariate itself); if 3, squares and cubes of each continuous covariate will be displayed, etc. If 1, the default, only the base covariate will be displayed. If int is numeric, poly will take on the value of int.

distance, distance.list
an optional formula or data frame containing distance values (e.g., propensity scores) or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look first in the argument to data, if specified, and next in the weightit or weightitMSM object. Note that propensity scores generated by weightit() and weightitMSM() are automatically included and named "prop.score". For weightitMSM objects, can be a list of distance values as described above, with one list entry per time period.

addl, addl.list
an optional formula or data frame containing additional covariates for which to present balance or a character vector containing their names. If a formula or variable names are specified, bal.tab() will look first in the argument to data, if specified, and next in the weightit or weightitMSM object. For weightitMSM objects, can be a list of additional covariate values as described above, with one list entry per time period.

data
an optional data frame containing variables that might be named in arguments to distance, addl, cluster, and imp. Can also be mids object, the output of a call to mice from the mice package, containing multiply imputed data sets. In this case, imp is automatically supplied using the imputation variable created from processing the mids object.

continuous
whether mean differences for continuous variables should be standardized ("std") or raw ("raw"). Default "std". Abbreviations allowed. This option can be set globally using set.cobalt.options. This argument is used to set std in col_w_smd and col_w_corr.

binary
whether mean differences for binary variables (i.e., difference in proportion) should be standardized ("std") or raw ("raw"). Default "raw". Abbreviations allowed. This option can be set globally using set.cobalt.options. This argument is used to set std in col_w_smd and col_w_corr.

s.d.denom
character; how the denominator for standardized mean differences should be calculated, if requested. See col_w_smd for allowable options. Can also be the name of the treatment group or "focal" if focal is specified in the original call to weightit. If not specified, bal.tab() will figure out which one is best based on the estimand of the weightit object: if ATT, "treated"; if ATC, "control", otherwise "pooled". Abbreviations allowed.

thresholds
a named vector of balance thresholds, where the name corresponds to the statistic (i.e., in stats) that the threshold applies to. For example, to request thresholds on mean differences and variance ratios, one can set thresholds = c(m = .05, v = 2). Requesting a threshold automatically requests the display of that statistic. See Details.

weights
a named list containing additional weights on which to assess balance. Each entry can be a vector of weights, the name of a variable in data that contains weights, or an object with a get.w method.

cluster
either a vector containing cluster membership for each unit or a string containing the name of the cluster membership variable in data or the weightit object. See bal.tab.cluster for details.

imp
either a vector containing imputation indices for each unit or a string containing the name of the imputation index variable in data or the original data set used in the call to weightit(). See bal.tab.imp for details. Not necessary if data is a mids object.

pairwise
when treatment is multi-category, whether balance should be computed for pairs of treatments or for each treatment against all groups combined. See bal.tab.multi for details.

abs
logical; whether displayed balance statistics should be in absolute value or not.

subset
a logical or numeric vector denoting whether each observation should be included or which observations should be included. If logical, it should be the same length as the variables in the original call to weightit() or weightitMSM(). NAs will be treated as FALSE. This can be used as an alternative to cluster to examine balance on subsets of the data.

quick
logical; if TRUE, will not compute any values that will not be displayed. Set to FALSE if computed values not displayed will be used later.

...
further arguments to control display of output. See display options for details.

Details

bal.tab.weightit() generates a list of balance summaries for the weightit object given.

The threshold argument controls whether extra columns should be inserted into the Balance table describing whether the balance statistics in question exceeded or were within the threshold. Including these thresholds also creates summary tables tallying the number of variables that exceeded and were within the threshold and displaying the variables with the greatest imbalance on that balance measure.

Value

For point treatments, if clusters and imputations are not specified, an object of class "bal.tab" containing balance summaries for the weightit object. See bal.tab for details.

If imputations are specified, an object of class "bal.tab.imp" containing balance summaries for each imputation and a summary of balance across imputations. See bal.tab.imp for details.

If weightit() is used with multi-category treatments, an object of class "bal.tab.multi" containing balance summaries for each pairwise treatment comparison. See bal.tab.multi for details.

If weightitMSM() is used for longitudinal treatments, an object of class "bal.tab.msm" containing balance summaries for each time period. See bal.tab.msm for details.

If clusters are specified, an object of class "bal.tab.cluster" containing balance summaries within each cluster and a summary of balance across clusters. See bal.tab.cluster for details.

See also

Examples

library(WeightIt)
data("lalonde", package = "cobalt")

## Basic propensity score weighting
w.out1 <- weightit(treat ~ age + educ + race +
                  married + nodegree + re74 + re75,
                  data = lalonde, method = "ps")
bal.tab(w.out1, un = TRUE, m.threshold = .1,
        v.threshold = 2)

#> Call
#>  weightit(formula = treat ~ age + educ + race + married + nodegree + 
#>     re74 + re75, data = lalonde, method = "ps")
#> 
#> Balance Measures
#>                 Type Diff.Un V.Ratio.Un Diff.Adj        M.Threshold V.Ratio.Adj
#> prop.score  Distance  1.7569     0.9211   0.1360                         0.9758
#> age          Contin. -0.2419     0.4400  -0.1676 Not Balanced, >0.1      0.3689
#> educ         Contin.  0.0448     0.4959   0.1296 Not Balanced, >0.1      0.5657
#> race_black    Binary  0.6404              0.0499     Balanced, <0.1            
#> race_hispan   Binary -0.0827              0.0047     Balanced, <0.1            
#> race_white    Binary -0.5577             -0.0546     Balanced, <0.1            
#> married       Binary -0.3236             -0.0944     Balanced, <0.1            
#> nodegree      Binary  0.1114             -0.0547     Balanced, <0.1            
#> re74         Contin. -0.5958     0.5181  -0.2740 Not Balanced, >0.1      0.8208
#> re75         Contin. -0.2870     0.9563  -0.1579 Not Balanced, >0.1      0.9562
#>                  V.Threshold
#> prop.score      Balanced, <2
#> age         Not Balanced, >2
#> educ            Balanced, <2
#> race_black                  
#> race_hispan                 
#> race_white                  
#> married                     
#> nodegree                    
#> re74            Balanced, <2
#> re75            Balanced, <2
#> 
#> Balance tally for mean differences
#>                    count
#> Balanced, <0.1         5
#> Not Balanced, >0.1     4
#> 
#> Variable with the greatest mean difference
#>  Variable Diff.Adj        M.Threshold
#>      re74   -0.274 Not Balanced, >0.1
#> 
#> Balance tally for variance ratios
#>                  count
#> Balanced, <2         4
#> Not Balanced, >2     1
#> 
#> Variable with the greatest variance ratio
#>  Variable V.Ratio.Adj      V.Threshold
#>       age      0.3689 Not Balanced, >2
#> 
#> Effective sample sizes
#>            Control Treated
#> Unadjusted  429.    185.  
#> Adjusted    329.01   58.33

## Entropy balancing for the ATE
w.out2 <- weightit(treat ~ age + educ + race +
                  married + nodegree + re74 + re75,
                  data = lalonde, method = "entropy",
                  estimand = "ATE")
bal.tab(w.out2)

#> Call
#>  weightit(formula = treat ~ age + educ + race + married + nodegree + 
#>     re74 + re75, data = lalonde, method = "entropy", estimand = "ATE")
#> 
#> Balance Measures
#>                Type Diff.Adj
#> age         Contin.  -0.0002
#> educ        Contin.   0.0001
#> race_black   Binary   0.0003
#> race_hispan  Binary  -0.0001
#> race_white   Binary  -0.0002
#> married      Binary  -0.0001
#> nodegree     Binary   0.0000
#> re74        Contin.  -0.0003
#> re75        Contin.  -0.0002
#> 
#> Effective sample sizes
#>            Control Treated
#> Unadjusted  429.     185. 
#> Adjusted    342.55    40.4

## IPW for longitudinal treatments
data("iptwExWide", package = "twang")
wmsm.out <- weightitMSM(list(tx1 ~ use0 + gender,
                             tx2 ~ use0 + gender + use1 + tx1,
                             tx3 ~ use0 + gender + use1 + tx1 + use2 + tx2),
                        data = iptwExWide,
                        stabilize = TRUE)
bal.tab(wmsm.out)

#> Call
#>  weightitMSM(formula.list = list(tx1 ~ use0 + gender, tx2 ~ use0 + 
#>     gender + use1 + tx1, tx3 ~ use0 + gender + use1 + tx1 + use2 + 
#>     tx2), data = iptwExWide, stabilize = TRUE)
#> 
#> Balance summary across all time points
#>              Times     Type Max.Diff.Adj
#> prop.score 1, 2, 3 Distance       0.3978
#> use0       1, 2, 3  Contin.       0.0672
#> gender     1, 2, 3   Binary       0.0102
#> use1          2, 3  Contin.       0.0286
#> tx1           2, 3   Binary       0.1602
#> use2             3  Contin.       0.0709
#> tx2              3   Binary       0.2345
#> 
#> Effective sample sizes
#>  - Time 1
#>            Control Treated
#> Unadjusted  294.    706.  
#> Adjusted    223.43  644.59
#>  - Time 2
#>            Control Treated
#> Unadjusted  492.    508.  
#> Adjusted    417.01  446.82
#>  - Time 3
#>            Control Treated
#> Unadjusted  415.    585.  
#> Adjusted    350.64  513.25