Lalonde's National Supported Work Demonstration Data

One of the datasets used by Dehejia and Wahba in their paper "Causal Effects in Non-Experimental Studies: Reevaluating the Evaluation of Training Programs." Versions of this data set have been used as an example data set in MatchIt, twang, Matching, and CBPS. The data set lalonde_mis is the same but with some values missing (set to NA).

data("lalonde")
data("lalonde_mis")

Format

A data frame with 614 observations on the following 9 variables.

treat

1 if treated in the National Supported Work Demonstration, 0 if from the Current Population Survey

age

age

educ

years of education

race

factor; black, Hispanic (hispan), or white

married

1 if married, 0 otherwise

nodegree

1 if no degree, 0 otherwise

re74

earnings in 1974 (pretreatment)

re75

earnings in 1975 (pretreatment)

re78

earnings in 1978 (outcome)

Source

https://users.nber.org/~rdehejia/data/.nswdata2.html

References

Lalonde, R. (1986). Evaluating the econometric evaluations of training programs with experimental data. American Economic Review 76: 604-620.

Dehejia, R.H. and Wahba, S. (1999). Causal Effects in Nonexperimental Studies: Re-Evaluating the Evaluation of Training Programs. Journal of the American Statistical Association 94: 1053-1062.